Split up the bloated BigMatrix.h file into separate files for each of the big matrix classes. Global methods and methods that involve multiple classes remain in BigMatrix.h.
