+++ SimTKcommon/BigMatrix/include/SimTKcommon/internal/BigMatrix.h
@@ -160,2148 +160,39 @@
 #include <limits>
 
 namespace SimTK {
+    template <class ELT>    class MatrixBase;
+    template <class ELT>    class VectorBase;
+    template <class ELT>    class RowVectorBase;
 
-template <class ELT>    class MatrixBase;
-template <class ELT>    class VectorBase;
-template <class ELT>    class RowVectorBase;
+    template <class ELT = Real> class MatrixView_;
+    template <class ELT = Real> class Matrix_;
 
-template <class ELT=Real> class MatrixView_;
-template <class ELT=Real> class Matrix_;
+    template <class ELT = Real> class VectorView_;
+    template <class ELT = Real> class Vector_;
 
-template <class ELT=Real> class VectorView_;
-template <class ELT=Real> class Vector_;
+    template <class ELT = Real> class RowVectorView_;
+    template <class ELT = Real> class RowVector_;
 
-template <class ELT=Real> class RowVectorView_;
-template <class ELT=Real> class RowVector_;
+    template <class ELT, class VECTOR_CLASS> class VectorIterator;
+}
 
-template <class ELT, class VECTOR_CLASS> class VectorIterator;
+#include "SimTKcommon/internal/MatrixBase.h"
+#include "SimTKcommon/internal/VectorBase.h"
+#include "SimTKcommon/internal/RowVectorBase.h"
 
-//  -------------------------------- MatrixBase --------------------------------
-/** @brief This is the common base class for Simbody's Vector_ and Matrix_
-classes for handling large, variable-sized vectors and matrices.
+#include "SimTKcommon/internal/MatrixView_.h"
+#include "SimTKcommon/internal/Matrix_.h"
 
-%MatrixBase does not normally appear in user programs. Instead classes
-Vector_ and Matrix_ are used, or more commonly the typedefs Vector and Matrix
-which are abbreviations for Vector_<Real> and Matrix_<Real>.
+#include "SimTKcommon/internal/VectorView_.h"
+#include "SimTKcommon/internal/Vector_.h"
 
-%Matrix base is a variable-size 2d matrix of Composite Numerical Type (CNT) 
-elements. This is a container of such elements, it is NOT a Composite Numerical
-Type itself. 
+#include "SimTKcommon/internal/RowVectorView_.h"
+#include "SimTKcommon/internal/RowVector_.h"
 
-<h2>Implementation</h2>
-MatrixBase<ELT> uses MatrixHelper<S> for implementation, where S 
-is ELT::Scalar, that is, the underlying float, double, long double,
-complex<float>, negator<conjugate<double>>, 
-etc. from which ELT is constructed. This is a finite set of which all
-members are explicitly instantiated in the implementation code, so 
-clients don't have to know how anything is implemented.
+#include "SimTKcommon/internal/VectorIterator.h"
 
-MatrixBase is the only class in the Matrix/Vector family which has any
-data members (it has exactly one MatrixHelper, which itself consists only
-of a single pointer to an opaque class). Thus all other objects
-in this family (that is, derived from MatrixBase) are exactly the same
-size in memory and may be "reinterpreted" as appropriate. For example,
-a Vector may be reinterpreted as a Matrix or vice versa, provided runtime
-requirements are met (e.g., exactly 1 column).
 
-Unlike the small matrix classes, very little is encoded in the type.
-Only the element type, and matrix vs. vector vs. row are in the type;
-everything else like shape, storage layout, and writability are handled
-at run time. **/
-//  ----------------------------------------------------------------------------
-template <class ELT> class MatrixBase {  
-public:
-    // These typedefs are handy, but despite appearances you cannot 
-    // treat a MatrixBase as a composite numerical type. That is,
-    // CNT<MatrixBase> will not compile, or if it does it won't be
-    // meaningful.
-
-    typedef ELT                                 E;
-    typedef typename CNT<E>::TNeg               ENeg;
-    typedef typename CNT<E>::TWithoutNegator    EWithoutNegator;
-    typedef typename CNT<E>::TReal              EReal;
-    typedef typename CNT<E>::TImag              EImag;
-    typedef typename CNT<E>::TComplex           EComplex;
-    typedef typename CNT<E>::THerm              EHerm;       
-    typedef typename CNT<E>::TPosTrans          EPosTrans;
-
-    typedef typename CNT<E>::TAbs               EAbs;
-    typedef typename CNT<E>::TStandard          EStandard;
-    typedef typename CNT<E>::TInvert            EInvert;
-    typedef typename CNT<E>::TNormalize         ENormalize;
-    typedef typename CNT<E>::TSqHermT           ESqHermT;
-    typedef typename CNT<E>::TSqTHerm           ESqTHerm;
-
-    typedef typename CNT<E>::Scalar             EScalar;
-    typedef typename CNT<E>::Number             ENumber;
-    typedef typename CNT<E>::StdNumber          EStdNumber;
-    typedef typename CNT<E>::Precision          EPrecision;
-    typedef typename CNT<E>::ScalarNormSq       EScalarNormSq;
-
-    typedef EScalar    Scalar;        // the underlying Scalar type
-    typedef ENumber    Number;        // negator removed from Scalar
-    typedef EStdNumber StdNumber;     // conjugate goes to complex
-    typedef EPrecision Precision;     // complex removed from StdNumber
-    typedef EScalarNormSq  ScalarNormSq;      // type of scalar^2
-
-    typedef MatrixBase<E>                T;
-    typedef MatrixBase<ENeg>             TNeg;
-    typedef MatrixBase<EWithoutNegator>  TWithoutNegator;
-    typedef MatrixBase<EReal>            TReal;
-    typedef MatrixBase<EImag>            TImag;
-    typedef MatrixBase<EComplex>         TComplex;
-    typedef MatrixBase<EHerm>            THerm; 
-    typedef MatrixBase<E>                TPosTrans;
-
-    typedef MatrixBase<EAbs>             TAbs;
-    typedef MatrixBase<EStandard>        TStandard;
-    typedef MatrixBase<EInvert>          TInvert;
-    typedef MatrixBase<ENormalize>       TNormalize;
-    typedef MatrixBase<ESqHermT>         TSqHermT;  // ~Mat*Mat
-    typedef MatrixBase<ESqTHerm>         TSqTHerm;  // Mat*~Mat
-
-    const MatrixCommitment& getCharacterCommitment() const {return helper.getCharacterCommitment();}
-    const MatrixCharacter& getMatrixCharacter()     const {return helper.getMatrixCharacter();}
-
-    /// Change the handle commitment for this matrix handle; only allowed if the 
-    /// handle is currently clear.
-    void commitTo(const MatrixCommitment& mc)
-    {   helper.commitTo(mc); }
-
-    // This gives the resulting matrix type when (m(i,j) op P) is applied to each element.
-    // It will have element types which are the regular composite result of E op P.
-    template <class P> struct EltResult { 
-        typedef MatrixBase<typename CNT<E>::template Result<P>::Mul> Mul;
-        typedef MatrixBase<typename CNT<E>::template Result<P>::Dvd> Dvd;
-        typedef MatrixBase<typename CNT<E>::template Result<P>::Add> Add;
-        typedef MatrixBase<typename CNT<E>::template Result<P>::Sub> Sub;
-    };
-
-    /// Return the number of rows m in the logical shape of this matrix.
-    int  nrow() const {return helper.nrow();}
-    /// Return the number of columns n in the logical shape of this matrix.
-    int  ncol() const {return helper.ncol();}
-
-    /// Return the number of elements in the \e logical shape of this matrix.
-    /// This has nothing to do with how many elements are actually stored;
-    /// it is simply the product of the logical number of rows and columns,
-    /// that is, nrow()*ncol(). Note that although each dimension is limited
-    /// to a 32 bit size, the product of those dimensions may be > 32 bits 
-    /// on a 64 bit machine so the return type may be larger than that of
-    /// nrow() and ncol().
-    ptrdiff_t nelt() const {return helper.nelt();}
-
-    /// Return true if either dimension of this Matrix is resizable.
-    bool isResizeable() const {return getCharacterCommitment().isResizeable();}
-
-    enum { 
-        NScalarsPerElement    = CNT<E>::NActualScalars,
-        CppNScalarsPerElement = sizeof(E) / sizeof(Scalar)
-    };
-  
-    /// The default constructor builds a 0x0 matrix managed by a helper that
-    /// understands how many scalars there are in one of our elements but is
-    /// otherwise uncommitted.
-    MatrixBase() : helper(NScalarsPerElement,CppNScalarsPerElement) {}
-
-    /// This constructor allocates the default matrix a completely uncommitted
-    /// matrix commitment, given particular initial dimensions.
-    MatrixBase(int m, int n) 
-    :   helper(NScalarsPerElement,CppNScalarsPerElement,MatrixCommitment(),m,n) {}
-
-    /// This constructor takes a handle commitment and allocates the default
-    /// matrix for that kind of commitment. If a dimension is set to a 
-    /// particular (unchangeable) value in the commitment then the initial
-    /// allocation will use that value. Unlocked dimensions are given the
-    /// smallest value consistent with other committed attributes, typically 0.
-    explicit MatrixBase(const MatrixCommitment& commitment) 
-    :   helper(NScalarsPerElement,CppNScalarsPerElement,commitment) {}
-
-
-    /// This constructor takes a handle commitment and allocates the default
-    /// matrix for that kind of commitment given particular initial minimum
-    /// dimensions, which cannot be larger than those permitted by the 
-    /// commitment.
-    MatrixBase(const MatrixCommitment& commitment, int m, int n) 
-    :   helper(NScalarsPerElement,CppNScalarsPerElement,commitment,m,n) {}
-
-    /// Copy constructor is a deep copy (not appropriate for views!).    
-    MatrixBase(const MatrixBase& b)
-      : helper(b.helper.getCharacterCommitment(), 
-               b.helper, typename MatrixHelper<Scalar>::DeepCopy()) { }
-
-    /// Implicit conversion from matrix with negated elements (otherwise this
-    /// is just like the copy constructor.
-    MatrixBase(const TNeg& b)
-      : helper(b.helper.getCharacterCommitment(),
-               b.helper, typename MatrixHelper<Scalar>::DeepCopy()) { }
-    
-    /// Copy assignment is a deep copy but behavior depends on type of lhs: if 
-    /// view, rhs must match. If owner, we reallocate and copy rhs.
-    MatrixBase& copyAssign(const MatrixBase& b) {
-        helper.copyAssign(b.helper);
-        return *this;
-    }
-    MatrixBase& operator=(const MatrixBase& b) { return copyAssign(b); }
-
-
-    /// View assignment is a shallow copy, meaning that we disconnect the MatrixBase 
-    /// from whatever it used to refer to (destructing as necessary), then make it a new view
-    /// for the data descriptor referenced by the source.
-    /// CAUTION: we always take the source as const, but that is ignored in 
-    /// determining whether the resulting view is writable. Instead, that is
-    /// inherited from the writability status of the source. We have to do this
-    /// in order to allow temporary view objects to be writable -- the compiler
-    /// creates temporaries like m(i,j,m,n) as const.
-    MatrixBase& viewAssign(const MatrixBase& src) {
-        helper.writableViewAssign(const_cast<MatrixHelper<Scalar>&>(src.helper));
-        return *this;
-    }
-
-    // default destructor
-
-    /// Initializing constructor with all of the initially-allocated elements
-    /// initialized to the same value. The given dimensions are treated as
-    /// minimum dimensions in case the commitment requires more. So it is 
-    /// always permissible to set them both to 0 in which case you'll get
-    /// the smallest matrix that satisfies the commitment, with each of its
-    /// elements (if any) set to the given initial value.
-    MatrixBase(const MatrixCommitment& commitment, int m, int n, const ELT& initialValue) 
-    :   helper(NScalarsPerElement, CppNScalarsPerElement, commitment, m, n)
-    {   helper.fillWith(reinterpret_cast<const Scalar*>(&initialValue)); }  
-
-    /// Initializing constructor with the initially-allocated elements
-    /// initialized from a C++ array of elements, which is provided in
-    /// <i>row major</i> order. The given dimensions are treated as
-    /// minimum dimensions in case the commitment requires more. The
-    /// array is presumed to be long enough to supply a value for each
-    /// element. Note that C++ packing for elements may be different than
-    /// Simmatrix packing of the same elements (Simmatrix packs them
-    /// more tightly in some cases). So you should not use this constructor
-    /// to copy elements from one Simmatrix matrix to another; this is
-    /// exclusively for initializing a Simmatrix from a C++ array.
-    MatrixBase(const MatrixCommitment& commitment, int m, int n, 
-               const ELT* cppInitialValuesByRow) 
-    :   helper(NScalarsPerElement, CppNScalarsPerElement, commitment, m, n)
-    {   helper.copyInByRowsFromCpp(reinterpret_cast<const Scalar*>(cppInitialValuesByRow)); }
-     
-    /// @name           Matrix view of pre-exising data
-    ///
-    /// Non-resizeable view of someone else's already-allocated 
-    /// memory of a size and storage type indicated by the supplied
-    /// MatrixCharacter. The \a spacing argument has different interpretations
-    /// depending on the storage format. Typically it is the leading
-    /// dimension for Lapack-style full storage or stride for a vector.
-    /// Spacing is in units like "number of scalars between elements" or
-    /// "number of scalars between columns" so it can be used to deal
-    /// with C++ packing vs. Simmatrix packing if necessary.
-    /// @{
-
-    /// Construct a read-only view of pre-existing data.
-    MatrixBase(const MatrixCommitment& commitment, 
-               const MatrixCharacter&  character, 
-               int spacing, const Scalar* data) // read only data
-    :   helper(NScalarsPerElement, CppNScalarsPerElement, 
-               commitment, character, spacing, data) {}  
-
-    /// Construct a writable view of pre-existing data.
-    MatrixBase(const MatrixCommitment& commitment, 
-               const MatrixCharacter&  character, 
-               int spacing, Scalar* data) // writable data
-    :   helper(NScalarsPerElement, CppNScalarsPerElement, 
-               commitment, character, spacing, data) {}  
-    /// @}
-        
-    // Create a new MatrixBase from an existing helper. Both shallow and deep copies are possible.
-    MatrixBase(const MatrixCommitment& commitment, 
-               MatrixHelper<Scalar>&   source, 
-               const typename MatrixHelper<Scalar>::ShallowCopy& shallow) 
-    :   helper(commitment, source, shallow) {}
-    MatrixBase(const MatrixCommitment&      commitment, 
-               const MatrixHelper<Scalar>&  source, 
-               const typename MatrixHelper<Scalar>::ShallowCopy& shallow) 
-    :   helper(commitment, source, shallow) {}
-    MatrixBase(const MatrixCommitment&      commitment, 
-               const MatrixHelper<Scalar>&  source, 
-               const typename MatrixHelper<Scalar>::DeepCopy& deep)    
-    :   helper(commitment, source, deep) {}
-
-    /// This restores the MatrixBase to the state it would be in had it
-    /// been constructed specifying only its handle commitment. The size will
-    /// have been reduced to the smallest size consistent with the commitment.
-    void clear() {helper.clear();}
-
-    MatrixBase& operator*=(const StdNumber& t)  { helper.scaleBy(t);              return *this; }
-    MatrixBase& operator/=(const StdNumber& t)  { helper.scaleBy(StdNumber(1)/t); return *this; }
-    MatrixBase& operator+=(const MatrixBase& r) { helper.addIn(r.helper);         return *this; }
-    MatrixBase& operator-=(const MatrixBase& r) { helper.subIn(r.helper);         return *this; }  
-
-    template <class EE> MatrixBase(const MatrixBase<EE>& b)
-      : helper(MatrixCommitment(),b.helper, typename MatrixHelper<Scalar>::DeepCopy()) { }
-
-    template <class EE> MatrixBase& operator=(const MatrixBase<EE>& b) 
-      { helper = b.helper; return *this; }
-    template <class EE> MatrixBase& operator+=(const MatrixBase<EE>& b) 
-      { helper.addIn(b.helper); return *this; }
-    template <class EE> MatrixBase& operator-=(const MatrixBase<EE>& b) 
-      { helper.subIn(b.helper); return *this; }
-
-    /// Matrix assignment to an element sets only the *diagonal* elements to
-    /// the indicated value; everything else is set to zero. This is particularly
-    /// useful for setting a Matrix to zero or to the identity; for other values
-    /// it creates a Matrix which acts like the scalar. That is, if the scalar
-    /// is s and we do M=s, then multiplying another Matrix B by the resulting 
-    /// diagonal matrix M gives the same result as multiplying B by s. That is
-    /// (M=s)*B == s*B.
-    ///
-    /// NOTE: this must be overridden for Vector and RowVector since then scalar
-    /// assignment is defined to copy the scalar to every element.
-    MatrixBase& operator=(const ELT& t) { 
-        setToZero(); updDiag().setTo(t); 
-        return *this;
-    }
-
-    /// Set M's diagonal elements to a "scalar" value S, and all off-diagonal
-    /// elements to zero. S can be any type which is assignable to an element
-    /// of type E. This is the same as the Matrix assignment operator M=S for
-    /// a scalar type S. It is overriden for Vector and Row types to behave
-    /// as elementwiseScalarAssign.
-    template <class S> inline MatrixBase&
-    scalarAssign(const S& s) {
-        setToZero(); updDiag().setTo(s);
-        return *this;
-    }
-
-    /// Add a scalar to M's diagonal. This is the same as the Matrix +=
-    /// operator. This is overridden for Vector and Row types to behave
-    /// as elementwiseAddScalarInPlace.
-    template <class S> inline MatrixBase&
-    scalarAddInPlace(const S& s) {
-        updDiag().elementwiseAddScalarInPlace(s);
-    }
-
-
-    /// Subtract a scalar from M's diagonal. This is the same as the Matrix -=
-    /// operator. This is overridden for Vector and Row types to behave
-    /// as elementwiseSubtractScalarInPlace.
-    template <class S> inline MatrixBase&
-    scalarSubtractInPlace(const S& s) {
-        updDiag().elementwiseSubtractScalarInPlace(s);
-    }
-
-    /// Set M(i,i) = S - M(i,i), M(i,j) = -M(i,j) for i!=j. This is overridden
-    /// for Vector and Row types to behave as elementwiseSubtractFromScalarInPlace.
-    template <class S> inline MatrixBase&
-    scalarSubtractFromLeftInPlace(const S& s) {
-        negateInPlace();
-        updDiag().elementwiseAddScalarInPlace(s); // yes, add
-    }
-
-    /// Set M(i,j) = M(i,j)*S for some "scalar" S. Actually S can be any
-    /// type for which E = E*S makes sense. That is, S must be conformant
-    /// with E and it must be possible to store the result back in an E.
-    /// This is the *= operator for M *= S and behaves the same way for
-    /// Matrix, Vector, and RowVector: every element gets multiplied in
-    /// place on the right by S.
-    template <class S> inline MatrixBase&
-    scalarMultiplyInPlace(const S&);
-
-    /// Set M(i,j) = S * M(i,j) for some "scalar" S. This is the same
-    /// as the above routine if S really is a scalar, but for S a more
-    /// complicated CNT it will be different.
-    template <class S> inline MatrixBase&
-    scalarMultiplyFromLeftInPlace(const S&);
-
-    /// Set M(i,j) = M(i,j)/S for some "scalar" S. Actually S can be any
-    /// type for which E = E/S makes sense. That is, S^-1 must be conformant
-    /// with E and it must be possible to store the result back in an E.
-    /// This is the /= operator for M /= S and behaves the same way for
-    /// Matrix, Vector, and RowVector: every element gets divided in
-    /// place on the right by S.
-    template <class S> inline MatrixBase&
-    scalarDivideInPlace(const S&);
-
-    /// Set M(i,j) = S/M(i,j) for some "scalar" S. Actually S can be any
-    /// type for which E = S/E makes sense. That is, S must be conformant
-    /// with E^-1 and it must be possible to store the result back in an E.
-    template <class S> inline MatrixBase&
-    scalarDivideFromLeftInPlace(const S&);
-
-
-    /// M = diag(r) * M; r must have nrow() elements.
-    /// That is, M[i] *= r[i].
-    template <class EE> inline MatrixBase& 
-    rowScaleInPlace(const VectorBase<EE>&);
-
-    /// Return type is a new matrix which will have the same dimensions as 'this' but
-    /// will have element types appropriate for the elementwise multiply being performed.
-    template <class EE> inline void 
-    rowScale(const VectorBase<EE>& r, typename EltResult<EE>::Mul& out) const;
-
-    template <class EE> inline typename EltResult<EE>::Mul 
-    rowScale(const VectorBase<EE>& r) const {
-        typename EltResult<EE>::Mul out(nrow(), ncol()); rowScale(r,out); return out;
-    }
-
-    /// M = M * diag(c); c must have ncol() elements.
-    /// That is, M(j) *= c[j].
-    template <class EE> inline MatrixBase& 
-    colScaleInPlace(const VectorBase<EE>&);
-
-	template <class EE> inline void 
-    colScale(const VectorBase<EE>& c, typename EltResult<EE>::Mul& out) const;
-
-	template <class EE> inline typename EltResult<EE>::Mul
-    colScale(const VectorBase<EE>& c) const {
-        typename EltResult<EE>::Mul out(nrow(), ncol()); colScale(c,out); return out;
-    }
-
-
-    /// M = diag(r) * M * diag(c); r must have nrow() elements;  must have ncol() elements.
-    /// That is, M(i,j) *= r[i]*c[j].
-    /// Having a combined row & column scaling operator means we can go through the matrix
-    /// memory once instead of twice.
-    template <class ER, class EC> inline MatrixBase& 
-    rowAndColScaleInPlace(const VectorBase<ER>& r, const VectorBase<EC>& c);
-
-    template <class ER, class EC> inline void 
-    rowAndColScale(const VectorBase<ER>& r, const VectorBase<EC>& c, 
-                   typename EltResult<typename VectorBase<ER>::template EltResult<EC>::Mul>::Mul& out) const;
-
-    template <class ER, class EC> inline typename EltResult<typename VectorBase<ER>::template EltResult<EC>::Mul>::Mul
-    rowAndColScale(const VectorBase<ER>& r, const VectorBase<EC>& c) const {
-        typename EltResult<typename VectorBase<ER>::template EltResult<EC>::Mul>::Mul 
-            out(nrow(), ncol()); 
-        rowAndColScale(r,c,out); return out;
-    }
-
-    /// Set M(i,j)=s for every element of M and some value s. This requires only
-    /// that s be assignment compatible with M's elements; s doesn't
-    /// actually have to be a scalar. Note that for Matrix types this behavior
-    /// is different than scalar assignment, which puts the scalar only on M's
-    /// diagonal and sets the rest of M to zero. For Vector and RowVector types,
-    /// this operator is identical to the normal assignment operator and
-    /// scalarAssignInPlace() method which also assign the scalar to every element.
-    template <class S> inline MatrixBase&
-    elementwiseAssign(const S& s);
-
-    /// Overloaded to allow an integer argument, which is converted to Real.
-    MatrixBase& elementwiseAssign(int s)
-    {   return elementwiseAssign<Real>(Real(s)); }
-
-    /// Set M(i,j) = M(i,j)^-1.
-    MatrixBase& elementwiseInvertInPlace();
-
-    void elementwiseInvert(MatrixBase<typename CNT<E>::TInvert>& out) const;
-
-    MatrixBase<typename CNT<E>::TInvert> elementwiseInvert() const {
-        MatrixBase<typename CNT<E>::TInvert> out(nrow(), ncol());
-        elementwiseInvert(out);
-        return out;
-    }
-
-    /// Set M(i,j)+=s for every element of M and some value s. This requires that s be
-    /// conformant with M's elements (of type E) and that the result can
-    /// be stored in an E. For Matrix types this behavior is different than
-    /// the normal += or scalarAddInPlace() operators, which add the scalar
-    /// only to the Matrix diagonal. For Vector and RowVector, this operator
-    /// is identical to += and scalarAddInPlace() which also add the scalar
-    /// to every element.
-    template <class S> inline MatrixBase&
-    elementwiseAddScalarInPlace(const S& s);
-
-    template <class S> inline void
-    elementwiseAddScalar(const S& s, typename EltResult<S>::Add&) const;
-
-    template <class S> inline typename EltResult<S>::Add
-    elementwiseAddScalar(const S& s) const {
-        typename EltResult<S>::Add out(nrow(), ncol());
-        elementwiseAddScalar(s,out);
-        return out;
-    }
-
-    /// Set M(i,j)-=s for every element of M and some value s. This requires that s be
-    /// conformant with M's elements (of type E) and that the result can
-    /// be stored in an E. For Matrix types this behavior is different than
-    /// the normal -= or scalarSubtractInPlace() operators, which subtract the scalar
-    /// only from the Matrix diagonal. For Vector and RowVector, this operator
-    /// is identical to -= and scalarSubtractInPlace() which also subtract the scalar
-    /// from every element.
-    template <class S> inline MatrixBase&
-    elementwiseSubtractScalarInPlace(const S& s);
-
-    template <class S> inline void
-    elementwiseSubtractScalar(const S& s, typename EltResult<S>::Sub&) const;
-
-    template <class S> inline typename EltResult<S>::Sub
-    elementwiseSubtractScalar(const S& s) const {
-        typename EltResult<S>::Sub out(nrow(), ncol());
-        elementwiseSubtractScalar(s,out);
-        return out;
-    }
-
-    /// Set M(i,j) = s - M(i,j) for every element of M and some value s. This requires that s be
-    /// conformant with M's elements (of type E) and that the result can
-    /// be stored in an E. For Matrix types this behavior is different than
-    /// the scalarSubtractFromLeftInPlace() operator, which subtracts only the diagonal
-    /// elements of M from s, while simply negating the off diagonal elements.
-    /// For Vector and RowVector, this operator
-    /// is identical to scalarSubtractFromLeftInPlace() which also subtracts every
-    /// element of M from the scalar.
-    template <class S> inline MatrixBase&
-    elementwiseSubtractFromScalarInPlace(const S& s);
-
-    template <class S> inline void
-    elementwiseSubtractFromScalar(
-        const S&, 
-        typename MatrixBase<S>::template EltResult<E>::Sub&) const;
-
-    template <class S> inline typename MatrixBase<S>::template EltResult<E>::Sub
-    elementwiseSubtractFromScalar(const S& s) const {
-        typename MatrixBase<S>::template EltResult<E>::Sub out(nrow(), ncol());
-        elementwiseSubtractFromScalar<S>(s,out);
-        return out;
-    }
-
-	/// M(i,j) *= R(i,j); R must have same dimensions as this.
-	template <class EE> inline MatrixBase& 
-    elementwiseMultiplyInPlace(const MatrixBase<EE>&);
-
-	template <class EE> inline void 
-    elementwiseMultiply(const MatrixBase<EE>&, typename EltResult<EE>::Mul&) const;
-
-	template <class EE> inline typename EltResult<EE>::Mul 
-    elementwiseMultiply(const MatrixBase<EE>& m) const {
-        typename EltResult<EE>::Mul out(nrow(), ncol()); 
-        elementwiseMultiply<EE>(m,out); 
-        return out;
-    }
-
-	/// M(i,j) = R(i,j) * M(i,j); R must have same dimensions as this.
-	template <class EE> inline MatrixBase& 
-    elementwiseMultiplyFromLeftInPlace(const MatrixBase<EE>&);
-
-	template <class EE> inline void 
-    elementwiseMultiplyFromLeft(
-        const MatrixBase<EE>&, 
-        typename MatrixBase<EE>::template EltResult<E>::Mul&) const;
-
-	template <class EE> inline typename MatrixBase<EE>::template EltResult<E>::Mul 
-    elementwiseMultiplyFromLeft(const MatrixBase<EE>& m) const {
-        typename EltResult<EE>::Mul out(nrow(), ncol()); 
-        elementwiseMultiplyFromLeft<EE>(m,out); 
-        return out;
-    }
-
-    /// M(i,j) /= R(i,j); R must have same dimensions as this.
-    template <class EE> inline MatrixBase& 
-    elementwiseDivideInPlace(const MatrixBase<EE>&);
-
-    template <class EE> inline void 
-    elementwiseDivide(const MatrixBase<EE>&, typename EltResult<EE>::Dvd&) const;
-
-    template <class EE> inline typename EltResult<EE>::Dvd 
-    elementwiseDivide(const MatrixBase<EE>& m) const {
-        typename EltResult<EE>::Dvd out(nrow(), ncol()); 
-        elementwiseDivide<EE>(m,out); 
-        return out;
-    }
-
-    /// M(i,j) = R(i,j) / M(i,j); R must have same dimensions as this.
-    template <class EE> inline MatrixBase& 
-    elementwiseDivideFromLeftInPlace(const MatrixBase<EE>&);
-
-    template <class EE> inline void 
-    elementwiseDivideFromLeft(
-        const MatrixBase<EE>&,
-        typename MatrixBase<EE>::template EltResult<E>::Dvd&) const;
-
-    template <class EE> inline typename MatrixBase<EE>::template EltResult<EE>::Dvd 
-    elementwiseDivideFromLeft(const MatrixBase<EE>& m) const {
-        typename MatrixBase<EE>::template EltResult<E>::Dvd out(nrow(), ncol()); 
-        elementwiseDivideFromLeft<EE>(m,out); 
-        return out;
-    }
-
-    /// Fill every element in current allocation with given element (or NaN or 0).
-    MatrixBase& setTo(const ELT& t) {helper.fillWith(reinterpret_cast<const Scalar*>(&t)); return *this;}
-    MatrixBase& setToNaN() {helper.fillWithScalar(CNT<StdNumber>::getNaN()); return *this;}
-    MatrixBase& setToZero() {helper.fillWithScalar(StdNumber(0)); return *this;}
- 
-    // View creating operators.   
-    inline RowVectorView_<ELT> row(int i) const;   // select a row
-    inline RowVectorView_<ELT> updRow(int i);
-    inline VectorView_<ELT>    col(int j) const;   // select a column
-    inline VectorView_<ELT>    updCol(int j);
-
-    RowVectorView_<ELT> operator[](int i) const {return row(i);}
-    RowVectorView_<ELT> operator[](int i)       {return updRow(i);}
-    VectorView_<ELT>    operator()(int j) const {return col(j);}
-    VectorView_<ELT>    operator()(int j)       {return updCol(j);}
-     
-    // Select a block.
-    inline MatrixView_<ELT> block(int i, int j, int m, int n) const;
-    inline MatrixView_<ELT> updBlock(int i, int j, int m, int n);
-
-    MatrixView_<ELT> operator()(int i, int j, int m, int n) const
-      { return block(i,j,m,n); }
-    MatrixView_<ELT> operator()(int i, int j, int m, int n)
-      { return updBlock(i,j,m,n); }
- 
-    // Hermitian transpose.
-    inline MatrixView_<EHerm> transpose() const;
-    inline MatrixView_<EHerm> updTranspose();
-
-    MatrixView_<EHerm> operator~() const {return transpose();}
-    MatrixView_<EHerm> operator~()       {return updTranspose();}
-
-    /// Select main diagonal (of largest leading square if rectangular) and
-    /// return it as a read-only view of the diagonal elements of this Matrix.
-    inline VectorView_<ELT> diag() const;
-    /// Select main diagonal (of largest leading square if rectangular) and
-    /// return it as a writable view of the diagonal elements of this Matrix.
-    inline VectorView_<ELT> updDiag();
-    /// This non-const version of diag() is an alternate name for updDiag()
-    /// available for historical reasons.
-    VectorView_<ELT> diag() {return updDiag();}
-
-    // Create a view of the real or imaginary elements. TODO
-    //inline MatrixView_<EReal> real() const;
-    //inline MatrixView_<EReal> updReal();
-    //inline MatrixView_<EImag> imag() const;
-    //inline MatrixView_<EImag> updImag();
-
-    // Overload "real" and "imag" for both read and write as a nod to convention. TODO
-    //MatrixView_<EReal> real() {return updReal();}
-    //MatrixView_<EReal> imag() {return updImag();}
-
-    // TODO: this routine seems ill-advised but I need it for the IVM port at the moment
-    TInvert invert() const {  // return a newly-allocated inverse; dump negator 
-        TInvert m(*this);
-        m.helper.invertInPlace();
-        return m;   // TODO - bad: makes an extra copy
-    }
-
-    void invertInPlace() {helper.invertInPlace();}
-
-    /// Matlab-compatible debug output.
-    void dump(const char* msg=0) const {
-        helper.dump(msg);
-    }
-
-
-    // This routine is useful for implementing friendlier Matrix expressions and operators.
-    // It maps closely to the Level-3 BLAS family of pxxmm() routines like sgemm(). The
-    // operation performed assumes that "this" is the result, and that "this" has 
-    // already been sized correctly to receive the result. We'll compute
-    //     this = beta*this + alpha*A*B
-    // If beta is 0 then "this" can be uninitialized. If alpha is 0 we promise not
-    // to look at A or B. The routine can work efficiently even if A and/or B are transposed
-    // by their views, so an expression like
-    //        C += s * ~A * ~B
-    // can be performed with the single equivalent call
-    //        C.matmul(1., s, Tr(A), Tr(B))
-    // where Tr(A) indicates a transposed view of the original A.
-    // The ultimate efficiency of this call depends on the data layout and views which
-    // are used for the three matrices.
-    // NOTE: neither A nor B can be the same matrix as 'this', nor views of the same data
-    // which would expose elements of 'this' that will be modified by this operation.
-    template <class ELT_A, class ELT_B>
-    MatrixBase& matmul(const StdNumber& beta,   // applied to 'this'
-                       const StdNumber& alpha, const MatrixBase<ELT_A>& A, const MatrixBase<ELT_B>& B)
-    {
-        helper.matmul(beta,alpha,A.helper,B.helper);
-        return *this;
-    }
-
-    /// Element selection for stored elements. These are the fastest element access
-    /// methods but may not be able to access all elements of the logical matrix when
-    /// some of its elements are not stored in memory. For example, a Hermitian matrix
-    /// stores only half its elements and other ones have to be calculated by conjugation
-    /// if they are to be returned as type ELT. (You can get them for free by recasting
-    /// the matrix so that the elements are reinterpreted as conjugates.) If you want
-    /// to guarantee that you can access the value of every element of a matrix, stored or not,
-    /// use getAnyElt() instead.
-    const ELT& getElt(int i, int j) const { return *reinterpret_cast<const ELT*>(helper.getElt(i,j)); }
-    ELT&       updElt(int i, int j)       { return *reinterpret_cast<      ELT*>(helper.updElt(i,j)); }
-
-    const ELT& operator()(int i, int j) const {return getElt(i,j);}
-    ELT&       operator()(int i, int j)       {return updElt(i,j);}
-
-    /// This returns a copy of the element value for any position in the logical matrix,
-    /// regardless of whether it is stored in memory. If necessary the element's value
-    /// is calculated. This is much slower than getElt() but less restrictive.
-    /// @see getElt()
-    void getAnyElt(int i, int j, ELT& value) const
-    {   helper.getAnyElt(i,j,reinterpret_cast<Scalar*>(&value)); }
-    ELT getAnyElt(int i, int j) const {ELT e; getAnyElt(i,j,e); return e;}
-
-    /// Scalar norm square is sum( squares of all scalars ). Note that this
-    /// is not very useful unless the elements are themselves scalars.
-    // TODO: very slow! Should be optimized at least for the case
-    //       where ELT is a Scalar.
-    ScalarNormSq scalarNormSqr() const {
-        const int nr=nrow(), nc=ncol();
-        ScalarNormSq sum(0);
-        for(int j=0;j<nc;++j) 
-            for (int i=0; i<nr; ++i)
-                sum += CNT<E>::scalarNormSqr((*this)(i,j));
-        return sum;
-    }
-
-    /// abs() is elementwise absolute value; that is, the return value has the same
-    /// dimension as this Matrix but with each element replaced by whatever it thinks
-    /// its absolute value is.
-    // TODO: very slow! Should be optimized at least for the case
-    //       where ELT is a Scalar.
-    void abs(TAbs& mabs) const {
-        const int nr=nrow(), nc=ncol();
-        mabs.resize(nr,nc);
-        for(int j=0;j<nc;++j) 
-            for (int i=0; i<nr; ++i)
-                mabs(i,j) = CNT<E>::abs((*this)(i,j));
-    }
-
-    /// abs() with the result as a function return. More convenient than the 
-    /// other abs() member function, but may involve an additional copy of the 
-    /// matrix.
-    TAbs abs() const { TAbs mabs; abs(mabs); return mabs; }
-
-    /// Return a Matrix of the same shape and contents as this one but
-    /// with the element type converted to one based on the standard
-    /// C++ scalar types: float, double, complex<float>,
-    /// complex<double>. That is, negator<>
-    /// and conjugate<> are eliminated from the element type by 
-    /// performing any needed negations computationally.
-    /// Note that this is actually producing a new matrix with new data;
-    /// you can also do this for free by reinterpreting the current
-    /// matrix as a different type, if you don't mind looking at
-    /// shared data.
-    TStandard standardize() const {
-        const int nr=nrow(), nc=ncol();
-        TStandard mstd(nr, nc);
-        for(int j=0;j<nc;++j) 
-            for (int i=0; i<nr; ++i)
-                mstd(i,j) = CNT<E>::standardize((*this)(i,j));
-        return mstd;
-    }
-
-    /// This is the scalar Frobenius norm, and its square. Note: if this is a 
-    /// Matrix then the Frobenius norm is NOT the same as the 2-norm, although
-    /// they are equivalent for Vectors.
-    ScalarNormSq normSqr() const { return scalarNormSqr(); }
-    // TODO -- not good; unnecessary overflow
-    typename CNT<ScalarNormSq>::TSqrt 
-        norm() const { return CNT<ScalarNormSq>::sqrt(scalarNormSqr()); }
-
-    /// We only allow RMS norm if the elements are scalars. If there are no 
-    /// elements in this Matrix, we'll define its RMS norm to be 0, although 
-    /// NaN might be a better choice.
-    typename CNT<ScalarNormSq>::TSqrt 
-    normRMS() const {
-        if (!CNT<ELT>::IsScalar)
-            SimTK_THROW1(Exception::Cant, "normRMS() only defined for scalar elements");
-        if (nelt() == 0)
-            return typename CNT<ScalarNormSq>::TSqrt(0);
-        return CNT<ScalarNormSq>::sqrt(scalarNormSqr()/nelt());
-    }
-
-    /// Form the column sums of this matrix, returned as a RowVector.
-    RowVector_<ELT> colSum() const {
-        const int cols = ncol();
-        RowVector_<ELT> row(cols);
-        for (int j = 0; j < cols; ++j)
-            helper.colSum(j, reinterpret_cast<Scalar*>(&row[j]));
-        return row;
-    }
-    /// Alternate name for colSum(); behaves like the Matlab function sum().
-    RowVector_<ELT> sum() const {return colSum();}
-
-    /// Form the row sums of this matrix, returned as a Vector.
-    Vector_<ELT> rowSum() const {
-        const int rows = nrow();
-        Vector_<ELT> col(rows);
-        for (int i = 0; i < rows; ++i)
-            helper.rowSum(i, reinterpret_cast<Scalar*>(&col[i]));
-        return col;
-    }
-
-    //TODO: make unary +/- return a self-view so they won't reallocate?
-    const MatrixBase& operator+() const {return *this; }
-    const TNeg&       negate()    const {return *reinterpret_cast<const TNeg*>(this); }
-    TNeg&             updNegate()       {return *reinterpret_cast<TNeg*>(this); }
-
-    const TNeg&       operator-() const {return negate();}
-    TNeg&             operator-()       {return updNegate();}
-
-    MatrixBase& negateInPlace() {(*this) *= EPrecision(-1); return *this;}
- 
-    /// Change the size of this matrix. This is only allowed for owner matrices. The
-    /// current storage format is retained, but all the data is lost. If you want
-    /// to keep the old data, use resizeKeep().
-    /// @see resizeKeep()
-    MatrixBase& resize(int m, int n)     { helper.resize(m,n); return *this; }
-    /// Change the size of this matrix, retaining as much of the old data as will
-    /// fit. This is only allowed for owner matrices. The
-    /// current storage format is retained, and the existing data is copied
-    /// into the new memory to the extent that it will fit.
-    /// @see resize()
-    MatrixBase& resizeKeep(int m, int n) { helper.resizeKeep(m,n); return *this; }
-
-    // This prevents shape changes in a Matrix that would otherwise allow it. No harm if is
-    // are called on a Matrix that is locked already; it always succeeds.
-    void lockShape() {helper.lockShape();}
-
-    // This allows shape changes again for a Matrix which was constructed to allow them
-    // but had them locked with the above routine. No harm if this is called on a Matrix
-    // that is already unlocked, but it is not allowed to call this on a Matrix which
-    // *never* allowed resizing. An exception will be thrown in that case.
-    void unlockShape() {helper.unlockShape();}
-    
-    // An assortment of handy conversions
-    const MatrixView_<ELT>& getAsMatrixView() const { return *reinterpret_cast<const MatrixView_<ELT>*>(this); }
-    MatrixView_<ELT>&       updAsMatrixView()       { return *reinterpret_cast<      MatrixView_<ELT>*>(this); } 
-    const Matrix_<ELT>&     getAsMatrix()     const { return *reinterpret_cast<const Matrix_<ELT>*>(this); }
-    Matrix_<ELT>&           updAsMatrix()           { return *reinterpret_cast<      Matrix_<ELT>*>(this); }
-         
-    const VectorView_<ELT>& getAsVectorView() const 
-      { assert(ncol()==1); return *reinterpret_cast<const VectorView_<ELT>*>(this); }
-    VectorView_<ELT>&       updAsVectorView()       
-      { assert(ncol()==1); return *reinterpret_cast<      VectorView_<ELT>*>(this); } 
-    const Vector_<ELT>&     getAsVector()     const 
-      { assert(ncol()==1); return *reinterpret_cast<const Vector_<ELT>*>(this); }
-    Vector_<ELT>&           updAsVector()           
-      { assert(ncol()==1); return *reinterpret_cast<      Vector_<ELT>*>(this); }
-    const VectorBase<ELT>& getAsVectorBase() const 
-      { assert(ncol()==1); return *reinterpret_cast<const VectorBase<ELT>*>(this); }
-    VectorBase<ELT>&       updAsVectorBase()       
-      { assert(ncol()==1); return *reinterpret_cast<      VectorBase<ELT>*>(this); } 
-                
-    const RowVectorView_<ELT>& getAsRowVectorView() const 
-      { assert(nrow()==1); return *reinterpret_cast<const RowVectorView_<ELT>*>(this); }
-    RowVectorView_<ELT>&       updAsRowVectorView()       
-      { assert(nrow()==1); return *reinterpret_cast<      RowVectorView_<ELT>*>(this); } 
-    const RowVector_<ELT>&     getAsRowVector()     const 
-      { assert(nrow()==1); return *reinterpret_cast<const RowVector_<ELT>*>(this); }
-    RowVector_<ELT>&           updAsRowVector()           
-      { assert(nrow()==1); return *reinterpret_cast<      RowVector_<ELT>*>(this); }        
-    const RowVectorBase<ELT>& getAsRowVectorBase() const 
-      { assert(nrow()==1); return *reinterpret_cast<const RowVectorBase<ELT>*>(this); }
-    RowVectorBase<ELT>&       updAsRowVectorBase()       
-      { assert(nrow()==1); return *reinterpret_cast<      RowVectorBase<ELT>*>(this); } 
-
-    // Access to raw data. We have to return the raw data
-    // pointer as pointer-to-scalar because we may pack the elements tighter
-    // than a C++ array would.
-
-    /// This is the number of consecutive scalars used to represent one
-    /// element of type ELT. This may be fewer than C++ would use for the
-    /// element, since it may introduce some padding.
-    int getNScalarsPerElement()  const {return NScalarsPerElement;}
-
-    /// This is like sizeof(ELT), but returning the number of bytes \e we use
-    /// to store the element which may be fewer than what C++ would use. We store
-    /// these packed elements adjacent to one another in memory.
-    int getPackedSizeofElement() const {return NScalarsPerElement*sizeof(Scalar);}
-
-    bool hasContiguousData() const {return helper.hasContiguousData();}
-    ptrdiff_t getContiguousScalarDataLength() const {
-        return helper.getContiguousDataLength();
-    }
-    const Scalar* getContiguousScalarData() const {
-        return helper.getContiguousData();
-    }
-    Scalar* updContiguousScalarData() {
-        return helper.updContiguousData();
-    }
-    void replaceContiguousScalarData(Scalar* newData, ptrdiff_t length, bool takeOwnership) {
-        helper.replaceContiguousData(newData,length,takeOwnership);
-    }
-    void replaceContiguousScalarData(const Scalar* newData, ptrdiff_t length) {
-        helper.replaceContiguousData(newData,length);
-    }
-    void swapOwnedContiguousScalarData(Scalar* newData, ptrdiff_t length, Scalar*& oldData) {
-        helper.swapOwnedContiguousData(newData,length,oldData);
-    }
-
-    /// Helper rep-stealing constructor. We take over ownership of this rep here. Note
-    /// that this \e defines the handle commitment for this handle. This is intended
-    /// for internal use only -- don't call this constructor unless you really 
-    /// know what you're doing.
-    explicit MatrixBase(MatrixHelperRep<Scalar>* hrep) : helper(hrep) {}
-
-protected:
-    const MatrixHelper<Scalar>& getHelper() const {return helper;}
-    MatrixHelper<Scalar>&       updHelper()       {return helper;}
-
-private:
-    MatrixHelper<Scalar> helper; // this is just one pointer
-
-    template <class EE> friend class MatrixBase;
-};
-
-
-
-//  -------------------------------- VectorBase --------------------------------
-/// This is a dataless rehash of the MatrixBase class to specialize it for Vectors.
-/// This mostly entails overriding a few of the methods. Note that all the MatrixBase
-/// operations remain available if you static_cast<> this up to a MatrixBase.
-//  ----------------------------------------------------------------------------
-template <class ELT> class VectorBase : public MatrixBase<ELT> {
-    typedef MatrixBase<ELT>                             Base;
-    typedef typename Base::ScalarNormSq                 ScalarNormSq;
-    typedef typename Base::EAbs                         EAbs;
-    typedef typename CNT<ELT>::Scalar                   Scalar;
-    typedef typename CNT<ELT>::Number                   Number;
-    typedef typename CNT<ELT>::StdNumber                StdNumber;
-    typedef VectorBase<ELT>                             T;
-    typedef VectorBase<typename CNT<ELT>::TAbs>         TAbs;
-    typedef VectorBase<typename CNT<ELT>::TNeg>         TNeg;
-    typedef RowVectorView_<typename CNT<ELT>::THerm>    THerm;
-public:  
-    //  ------------------------------------------------------------------------
-    /// @name       VectorBase "owner" construction
-    ///
-    /// These constructors create new VectorBase objects which own their
-    /// own data and are (at least by default) resizable. The resulting matrices
-    /// are m X 1 with the number of columns locked at 1. If there is any data
-    /// allocated but not explicitly initialized, that data will be uninitialized
-    /// garbage in Release builds but will be initialized to NaN (at a performance
-    /// cost) in Debug builds.
-    /// @{
-
-    /// Default constructor makes a 0x1 matrix locked at 1 column; you can
-    /// provide an initial allocation if you want.
-    explicit VectorBase(int m=0) : Base(MatrixCommitment::Vector(), m, 1) {}
-
-    /// Copy constructor is a deep copy (not appropriate for views!). That
-    /// means it creates a new, densely packed vector whose elements are
-    /// initialized from the source object.
-    VectorBase(const VectorBase& source) : Base(source) {}
-
-    /// Implicit conversion from compatible vector with negated elements.
-    VectorBase(const TNeg& source) : Base(source) {}
-
-    /// Construct an owner vector of length m, with each element initialized to
-    /// the given value.
-    VectorBase(int m, const ELT& initialValue)
-    :   Base(MatrixCommitment::Vector(),m,1,initialValue) {}  
-
-    /// Construct an owner vector of length m, with the elements initialized sequentially
-    /// from a C++ array of elements which is assumed to be of length m. Note that we
-    /// are expecting C++ packing; don't use this to initialize one Simmatrix vector
-    /// from another because Simmatrix may pack its elements more densely than C++.
-    VectorBase(int m, const ELT* cppInitialValues)
-    :   Base(MatrixCommitment::Vector(),m,1,cppInitialValues) {}
-    /// @}
-
-    //  ------------------------------------------------------------------------
-    /// @name       VectorBase construction from pre-existing data
-    ///
-    /// Construct a non-resizeable, VectorBase view of externally supplied data. Note that
-    /// stride should be interpreted as "the number of scalars between elements" and
-    /// for composite elements may have a different value if the source is a C++ array 
-    /// of elements vs. a Simmatrix packed data array. We provide constructors for
-    /// both read-only and writable external data.
-    /// @{
-
-    /// Construct a read-only view of existing data.
-    VectorBase(int m, int stride, const Scalar* s)
-    :   Base(MatrixCommitment::Vector(m), MatrixCharacter::Vector(m),stride,s) { }
-    /// Construct a writable view into existing data.
-    VectorBase(int m, int stride, Scalar* s)
-    :   Base(MatrixCommitment::Vector(m), MatrixCharacter::Vector(m),stride,s) { }
-    /// @}
-        
-    //  ------------------------------------------------------------------------
-    /// @name       VectorBase construction from an existing Helper.
-    ///
-    /// Create a new VectorBase from an existing helper. Both shallow (view) and deep 
-    /// copies are possible. For shallow copies, there is a constructor providing a read-only
-    /// view of the original data and one providing a writable view into the original data.
-    /// @{
-
-    /// Construct a writable view into the source data.
-    VectorBase(MatrixHelper<Scalar>& h, const typename MatrixHelper<Scalar>::ShallowCopy& s) 
-    :   Base(MatrixCommitment::Vector(), h,s) { }
-    /// Construct a read-only view of the source data.
-    VectorBase(const MatrixHelper<Scalar>& h, const typename MatrixHelper<Scalar>::ShallowCopy& s) 
-    :   Base(MatrixCommitment::Vector(), h,s) { }
-    /// Construct a new owner vector initialized with the data from the source.
-    VectorBase(const MatrixHelper<Scalar>& h, const typename MatrixHelper<Scalar>::DeepCopy& d)    
-    :   Base(MatrixCommitment::Vector(), h,d) { }
-    /// @}
-
-    // This gives the resulting vector type when (v[i] op P) is applied to each element.
-    // It will have element types which are the regular composite result of ELT op P.
-    template <class P> struct EltResult { 
-        typedef VectorBase<typename CNT<ELT>::template Result<P>::Mul> Mul;
-        typedef VectorBase<typename CNT<ELT>::template Result<P>::Dvd> Dvd;
-        typedef VectorBase<typename CNT<ELT>::template Result<P>::Add> Add;
-        typedef VectorBase<typename CNT<ELT>::template Result<P>::Sub> Sub;
-    };
-
-    /// Copy assignment is deep copy but behavior depends on type of lhs: if view, rhs
-    /// must match. If owner, we reallocate and copy rhs.
-    VectorBase& operator=(const VectorBase& b) {
-        Base::operator=(b); return *this;
-    }
-
-    // default destructor
-
-
-    VectorBase& operator*=(const StdNumber& t)  { Base::operator*=(t); return *this; }
-    VectorBase& operator/=(const StdNumber& t)  { Base::operator/=(t); return *this; }
-    VectorBase& operator+=(const VectorBase& r) { Base::operator+=(r); return *this; }
-    VectorBase& operator-=(const VectorBase& r) { Base::operator-=(r); return *this; }  
-
-
-    template <class EE> VectorBase& operator=(const VectorBase<EE>& b) 
-      { Base::operator=(b);  return *this; } 
-    template <class EE> VectorBase& operator+=(const VectorBase<EE>& b) 
-      { Base::operator+=(b); return *this; } 
-    template <class EE> VectorBase& operator-=(const VectorBase<EE>& b) 
-      { Base::operator-=(b); return *this; } 
-
-
-    /// Fill current allocation with copies of element. Note that this is not the 
-    /// same behavior as assignment for Matrices, where only the diagonal is set (and
-    /// everything else is set to zero.)
-    VectorBase& operator=(const ELT& t) { Base::setTo(t); return *this; }  
-
-    /// There's only one column here so it's a bit weird to use rowScale rather than
-    /// elementwiseMultiply, but there's nothing really wrong with it. Using colScale
-    /// would be really wacky since it is the same as a scalar multiply. We won't support
-    /// colScale here except through inheritance where it will not be much use.
-    template <class EE> VectorBase& rowScaleInPlace(const VectorBase<EE>& v)
-	{ Base::template rowScaleInPlace<EE>(v); return *this; }
-    template <class EE> inline void rowScale(const VectorBase<EE>& v, typename EltResult<EE>::Mul& out) const
-	{ Base::rowScale(v,out); }
-    template <class EE> inline typename EltResult<EE>::Mul rowScale(const VectorBase<EE>& v) const
-	{ typename EltResult<EE>::Mul out(nrow()); Base::rowScale(v,out); return out; }
-
-    /** Return the root-mean-square (RMS) norm of a Vector of scalars, with 
-    optional return of the index of the element of largest absolute value. 
-    The RMS norm of a Vector v of length n is rms=sqrt(~v*v/n). If n==0 we
-    define the RMS norm to be zero but return the element index as -1. **/
-    typename CNT<ScalarNormSq>::TSqrt 
-    normRMS(int* worstOne=0) const {
-        if (!CNT<ELT>::IsScalar)
-            SimTK_THROW1(Exception::Cant, 
-                "Vector::normRMS() only defined for scalar elements.");
-        const int n = nrow();
-        if (n == 0) {
-            if (worstOne) *worstOne = -1;
-            return typename CNT<ScalarNormSq>::TSqrt(0);
-        }
-
-        ScalarNormSq sumsq = 0;
-        if (worstOne) {
-            *worstOne = 0;
-            ScalarNormSq maxsq = 0; 
-            for (int i=0; i<n; ++i) {
-                const ScalarNormSq v2 = square((*this)[i]);
-                if (v2 > maxsq) maxsq=v2, *worstOne=i;
-                sumsq += v2;
-            }
-        } else { // don't track the worst element
-            for (int i=0; i<n; ++i) {
-                const ScalarNormSq v2 = square((*this)[i]);
-                sumsq += v2;
-            }
-        }
-
-        return CNT<ScalarNormSq>::sqrt(sumsq/n);
-    }
-
-    /** Return the weighted root-mean-square (WRMS) norm of a Vector of 
-    scalars, with optional return of the index of the weighted element of 
-    largest absolute value. The WRMS norm of a Vector v of length n with
-    weights w is wrms=sqrt(sum_i((w_i*v_i)^2))/n). If n==0 we
-    define the WRMS norm to be zero but return the element index as -1. **/
-    template <class EE>
-    typename CNT<ScalarNormSq>::TSqrt 
-    weightedNormRMS(const VectorBase<EE>& w, int* worstOne=0) const {
-        if (!CNT<ELT>::IsScalar || !CNT<EE>::IsScalar)
-            SimTK_THROW1(Exception::Cant, 
-            "Vector::weightedNormRMS() only defined for scalar elements"
-            " and weights.");
-        const int n = nrow();
-        assert(w.nrow()==n);
-        if (n == 0) {
-            if (worstOne) *worstOne = -1;
-            return typename CNT<ScalarNormSq>::TSqrt(0);
-        }
-
-        ScalarNormSq sumsq = 0;
-        if (worstOne) {
-            *worstOne = 0;
-            ScalarNormSq maxsq = 0; 
-            for (int i=0; i<n; ++i) {
-                const ScalarNormSq wv2 = square(w[i]*(*this)[i]);
-                if (wv2 > maxsq) maxsq=wv2, *worstOne=i;
-                sumsq += wv2;
-            }
-        } else { // don't track the worst element
-            for (int i=0; i<n; ++i) {
-                const ScalarNormSq wv2 = square(w[i]*(*this)[i]);
-                sumsq += wv2;
-            }
-        }
-
-        return CNT<ScalarNormSq>::sqrt(sumsq/n);
-    }
-
-    /** Return the infinity norm (max absolute value) of a Vector of scalars, 
-    with optional return of the index of the element of largest absolute value. 
-    The Inf norm of a Vector v is inf=max_i(|v_i|). If n==0 we
-    define the Inf norm to be zero but return the element index as -1. **/
-    EAbs normInf(int* worstOne=0) const {
-        if (!CNT<ELT>::IsScalar)
-            SimTK_THROW1(Exception::Cant, 
-                "Vector::normInf() only defined for scalar elements.");
-        const int n = nrow();
-        if (n == 0) {
-            if (worstOne) *worstOne = -1;
-            return EAbs(0);
-        }
-
-        EAbs maxabs = 0;
-        if (worstOne) {
-            *worstOne = 0;
-            for (int i=0; i<n; ++i) {
-                const EAbs a = std::abs((*this)[i]);
-                if (a > maxabs) maxabs=a, *worstOne=i;
-            }
-        } else { // don't track the worst element
-            for (int i=0; i<n; ++i) {
-                const EAbs a = std::abs((*this)[i]);
-                if (a > maxabs) maxabs=a;
-            }
-        }
-
-        return maxabs;
-    }
-
-    /** Return the weighted infinity norm (max absolute value) WInf of a Vector
-    of scalars, with optional return of the index of the weighted element of 
-    largest absolute value. The WInf norm of a Vector v of length n with
-    weights w is winf=max_i(|w_i*v_i|). If n==0 we
-    define the WInf norm to be zero but return the element index as -1. **/
-    template <class EE>
-    EAbs weightedNormInf(const VectorBase<EE>& w, int* worstOne=0) const {
-        if (!CNT<ELT>::IsScalar || !CNT<EE>::IsScalar)
-            SimTK_THROW1(Exception::Cant, 
-            "Vector::weightedNormInf() only defined for scalar elements"
-            " and weights.");
-        const int n = nrow();
-        assert(w.nrow()==n);
-        if (n == 0) {
-            if (worstOne) *worstOne = -1;
-            return EAbs(0);
-        }
-
-        EAbs maxabs = 0;
-        if (worstOne) {
-            *worstOne = 0;
-            for (int i=0; i<n; ++i) {
-                const EAbs wv = std::abs(w[i]*(*this)[i]);
-                if (wv > maxabs) maxabs=wv, *worstOne=i;
-            }
-        } else { // don't track the worst element
-            for (int i=0; i<n; ++i) {
-                const EAbs wv = std::abs(w[i]*(*this)[i]);
-                if (wv > maxabs) maxabs=wv;
-            }
-        }
-
-        return maxabs;
-    }
-
-    /// Set this[i] = this[i]^-1.
-    VectorBase& elementwiseInvertInPlace() {
-        Base::elementwiseInvertInPlace();
-        return *this;
-    }
-
-    /// Set supplied out[i] = this[i]^-1
-    void elementwiseInvert(VectorBase<typename CNT<ELT>::TInvert>& out) const {
-        Base::elementwiseInvert(out);
-    }
-
-    /// Return out[i]=this[i]^-1 as function return.
-    VectorBase<typename CNT<ELT>::TInvert> elementwiseInvert() const {
-        VectorBase<typename CNT<ELT>::TInvert> out(nrow());
-        Base::elementwiseInvert(out);
-        return out;
-    }
-
-    // elementwise multiply
-    template <class EE> VectorBase& elementwiseMultiplyInPlace(const VectorBase<EE>& r)
-	{ Base::template elementwiseMultiplyInPlace<EE>(r); return *this; }
-    template <class EE> inline void elementwiseMultiply(const VectorBase<EE>& v, typename EltResult<EE>::Mul& out) const
-	{ Base::template elementwiseMultiply<EE>(v,out); }
-    template <class EE> inline typename EltResult<EE>::Mul elementwiseMultiply(const VectorBase<EE>& v) const
-	{ typename EltResult<EE>::Mul out(nrow()); Base::template elementwiseMultiply<EE>(v,out); return out; }
-
-    // elementwise multiply from left
-    template <class EE> VectorBase& elementwiseMultiplyFromLeftInPlace(const VectorBase<EE>& r)
-	{ Base::template elementwiseMultiplyFromLeftInPlace<EE>(r); return *this; }
-    template <class EE> inline void 
-    elementwiseMultiplyFromLeft(
-        const VectorBase<EE>& v, 
-        typename VectorBase<EE>::template EltResult<ELT>::Mul& out) const
-	{ 
-        Base::template elementwiseMultiplyFromLeft<EE>(v,out);
-    }
-	template <class EE> inline typename VectorBase<EE>::template EltResult<ELT>::Mul 
-    elementwiseMultiplyFromLeft(const VectorBase<EE>& v) const
-	{ 
-        typename VectorBase<EE>::template EltResult<ELT>::Mul out(nrow()); 
-        Base::template elementwiseMultiplyFromLeft<EE>(v,out); 
-        return out;
-    }
-
-    // elementwise divide
-    template <class EE> VectorBase& elementwiseDivideInPlace(const VectorBase<EE>& r)
-	{ Base::template elementwiseDivideInPlace<EE>(r); return *this; }
-    template <class EE> inline void elementwiseDivide(const VectorBase<EE>& v, typename EltResult<EE>::Dvd& out) const
-	{ Base::template elementwiseDivide<EE>(v,out); }
-    template <class EE> inline typename EltResult<EE>::Dvd elementwiseDivide(const VectorBase<EE>& v) const
-	{ typename EltResult<EE>::Dvd out(nrow()); Base::template elementwiseDivide<EE>(v,out); return out; }
-
-    // elementwise divide from left
-    template <class EE> VectorBase& elementwiseDivideFromLeftInPlace(const VectorBase<EE>& r)
-	{ Base::template elementwiseDivideFromLeftInPlace<EE>(r); return *this; }
-    template <class EE> inline void 
-    elementwiseDivideFromLeft(
-        const VectorBase<EE>& v, 
-        typename VectorBase<EE>::template EltResult<ELT>::Dvd& out) const
-	{ 
-        Base::template elementwiseDivideFromLeft<EE>(v,out);
-    }
-	template <class EE> inline typename VectorBase<EE>::template EltResult<ELT>::Dvd 
-    elementwiseDivideFromLeft(const VectorBase<EE>& v) const
-	{ 
-        typename VectorBase<EE>::template EltResult<ELT>::Dvd out(nrow()); 
-        Base::template elementwiseDivideFromLeft<EE>(v,out); 
-        return out;
-    }
-
-    // Implicit conversions are allowed to Vector or Matrix, but not to RowVector.   
-    operator const Vector_<ELT>&()     const { return *reinterpret_cast<const Vector_<ELT>*>(this); }
-    operator       Vector_<ELT>&()           { return *reinterpret_cast<      Vector_<ELT>*>(this); }
-    operator const VectorView_<ELT>&() const { return *reinterpret_cast<const VectorView_<ELT>*>(this); }
-    operator       VectorView_<ELT>&()       { return *reinterpret_cast<      VectorView_<ELT>*>(this); }
-    
-    operator const Matrix_<ELT>&()     const { return *reinterpret_cast<const Matrix_<ELT>*>(this); }
-    operator       Matrix_<ELT>&()           { return *reinterpret_cast<      Matrix_<ELT>*>(this); } 
-    operator const MatrixView_<ELT>&() const { return *reinterpret_cast<const MatrixView_<ELT>*>(this); }
-    operator       MatrixView_<ELT>&()       { return *reinterpret_cast<      MatrixView_<ELT>*>(this); } 
-
-
-    // size() for Vectors is Base::nelt() but returns int instead of ptrdiff_t.
-    int size() const { 
-	assert(Base::nelt() <= (ptrdiff_t)std::numeric_limits<int>::max()); 
-	assert(Base::ncol()==1);
-	return (int)Base::nelt();
-    }
-    int       nrow() const {assert(Base::ncol()==1); return Base::nrow();}
-    int       ncol() const {assert(Base::ncol()==1); return Base::ncol();}
-    ptrdiff_t nelt() const {assert(Base::ncol()==1); return Base::nelt();}
-
-    // Override MatrixBase operators to return the right shape
-    TAbs abs() const {TAbs result; Base::abs(result); return result;}
-    
-    // Override MatrixBase indexing operators          
-    const ELT& operator[](int i) const {return *reinterpret_cast<const ELT*>(Base::getHelper().getElt(i));}
-    ELT&       operator[](int i)       {return *reinterpret_cast<ELT*>      (Base::updHelper().updElt(i));}
-    const ELT& operator()(int i) const {return *reinterpret_cast<const ELT*>(Base::getHelper().getElt(i));}
-    ELT&       operator()(int i)       {return *reinterpret_cast<ELT*>      (Base::updHelper().updElt(i));}
-         
-    // Block (contiguous subvector) view creation      
-    VectorView_<ELT> operator()(int i, int m) const {return Base::operator()(i,0,m,1).getAsVectorView();}
-    VectorView_<ELT> operator()(int i, int m)       {return Base::operator()(i,0,m,1).updAsVectorView();}
-
-    // Indexed view creation (arbitrary subvector). Indices must be 
-    // monotonically increasing.
-    VectorView_<ELT> index(const Array_<int>& indices) const {
-        MatrixHelper<Scalar> h(Base::getHelper().getCharacterCommitment(), 
-                               Base::getHelper(), indices);
-        return VectorView_<ELT>(h);
-    }
-    VectorView_<ELT> updIndex(const Array_<int>& indices) {
-        MatrixHelper<Scalar> h(Base::getHelper().getCharacterCommitment(), 
-                               Base::updHelper(), indices);
-        return VectorView_<ELT>(h);
-    }
-
-    VectorView_<ELT> operator()(const Array_<int>& indices) const {return index(indices);}
-    VectorView_<ELT> operator()(const Array_<int>& indices)       {return updIndex(indices);}
- 
-    // Hermitian transpose.
-    THerm transpose() const {return Base::transpose().getAsRowVectorView();}
-    THerm updTranspose()    {return Base::updTranspose().updAsRowVectorView();}
-
-    THerm operator~() const {return transpose();}
-    THerm operator~()       {return updTranspose();}
-
-    const VectorBase& operator+() const {return *this; }
-
-    // Negation
-
-    const TNeg& negate()    const {return *reinterpret_cast<const TNeg*>(this); }
-    TNeg&       updNegate()       {return *reinterpret_cast<TNeg*>(this); }
-
-    const TNeg& operator-() const {return negate();}
-    TNeg&       operator-()       {return updNegate();}
-
-    VectorBase& resize(int m)     {Base::resize(m,1); return *this;}
-    VectorBase& resizeKeep(int m) {Base::resizeKeep(m,1); return *this;}
-
-    //TODO: this is not re-locking the number of columns at 1.
-    void clear() {Base::clear(); Base::resize(0,1);}
-
-    ELT sum() const {ELT s; Base::getHelper().sum(reinterpret_cast<Scalar*>(&s)); return s; } // add all the elements        
-    VectorIterator<ELT, VectorBase<ELT> > begin() {
-        return VectorIterator<ELT, VectorBase<ELT> >(*this, 0);
-    }
-    VectorIterator<ELT, VectorBase<ELT> > end() {
-        return VectorIterator<ELT, VectorBase<ELT> >(*this, size());
-    }
-
-protected:
-    // Create a VectorBase handle using a given helper rep. 
-    explicit VectorBase(MatrixHelperRep<Scalar>* hrep) : Base(hrep) {}
-
-private:
-    // NO DATA MEMBERS ALLOWED
-};
-
-
-
-//  ------------------------------- RowVectorBase ------------------------------
-/// This is a dataless rehash of the MatrixBase class to specialize it for RowVectors.
-/// This mostly entails overriding a few of the methods. Note that all the MatrixBase
-/// operations remain available if you static_cast<> this up to a MatrixBase.
-//  ----------------------------------------------------------------------------
-template <class ELT> class RowVectorBase : public MatrixBase<ELT> {
-    typedef MatrixBase<ELT>                             Base;
-    typedef typename CNT<ELT>::Scalar                   Scalar;
-    typedef typename CNT<ELT>::Number                   Number;
-    typedef typename CNT<ELT>::StdNumber                StdNumber;
-    typedef RowVectorBase<ELT>                          T;
-    typedef RowVectorBase<typename CNT<ELT>::TAbs>      TAbs;
-    typedef RowVectorBase<typename CNT<ELT>::TNeg>      TNeg;
-    typedef VectorView_<typename CNT<ELT>::THerm>       THerm;
-public: 
-    //  ------------------------------------------------------------------------
-    /// @name       RowVectorBase "owner" construction
-    ///
-    /// These constructors create new RowVectorBase objects which own their
-    /// own data and are (at least by default) resizable. The resulting matrices
-    /// are 1 x n with the number of rows locked at 1. If there is any data
-    /// allocated but not explicitly initialized, that data will be uninitialized
-    /// garbage in Release builds but will be initialized to NaN (at a performance
-    /// cost) in Debug builds.
-    /// @{
-
-    /// Default constructor makes a 1x0 matrix locked at 1 row; you can
-    /// provide an initial allocation if you want.
-    explicit RowVectorBase(int n=0) : Base(MatrixCommitment::RowVector(), 1, n) {}
-    
-    /// Copy constructor is a deep copy (not appropriate for views!). That
-    /// means it creates a new, densely packed vector whose elements are
-    /// initialized from the source object.    
-    RowVectorBase(const RowVectorBase& source) : Base(source) {}
-
-    /// Implicit conversion from compatible row vector with negated elements.
-    RowVectorBase(const TNeg& source) : Base(source) {}
-
-    /// Construct an owner row vector of length n, with each element initialized to
-    /// the given value.
-    RowVectorBase(int n, const ELT& initialValue)
-    :   Base(MatrixCommitment::RowVector(),1,n,initialValue) {}  
-
-    /// Construct an owner vector of length n, with the elements initialized sequentially
-    /// from a C++ array of elements which is assumed to be of length n. Note that we
-    /// are expecting C++ packing; don't use this to initialize one Simmatrix vector
-    /// from another because Simmatrix may pack its elements more densely than C++.
-    RowVectorBase(int n, const ELT* cppInitialValues)
-    :   Base(MatrixCommitment::RowVector(),1,n,cppInitialValues) {}
-    /// @}
-
-    //  ------------------------------------------------------------------------
-    /// @name       RowVectorBase construction from pre-existing data
-    ///
-    /// Construct a non-resizeable, RowVectorBase view of externally supplied data. Note that
-    /// stride should be interpreted as "the number of scalars between elements" and
-    /// for composite elements may have a different value if the source is a C++ array 
-    /// of elements vs. a Simmatrix packed data array. We provide constructors for
-    /// both read-only and writable external data.
-    /// @{
-
-    /// Construct a read-only view of existing data.
-    RowVectorBase(int n, int stride, const Scalar* s)
-    :   Base(MatrixCommitment::RowVector(n), MatrixCharacter::RowVector(n),stride,s) { }
-    /// Construct a writable view into existing data.
-    RowVectorBase(int n, int stride, Scalar* s)
-    :   Base(MatrixCommitment::RowVector(n), MatrixCharacter::RowVector(n),stride,s) { }
-    /// @}
-
-    //  ------------------------------------------------------------------------
-    /// @name       RowVectorBase construction from an existing Helper.
-    ///
-    /// Create a new RowVectorBase from an existing helper. Both shallow (view) and deep 
-    /// copies are possible. For shallow copies, there is a constructor providing a read-only
-    /// view of the original data and one providing a writable view into the original data.
-    /// @{
-
-    /// Construct a writable view into the source data.
-    RowVectorBase(MatrixHelper<Scalar>& h, const typename MatrixHelper<Scalar>::ShallowCopy& s) 
-    :   Base(MatrixCommitment::RowVector(), h,s) { }
-    /// Construct a read-only view of the source data.
-    RowVectorBase(const MatrixHelper<Scalar>& h, const typename MatrixHelper<Scalar>::ShallowCopy& s) 
-    :   Base(MatrixCommitment::RowVector(), h,s) { }
-    /// Construct a new owner vector initialized with the data from the source.
-    RowVectorBase(const MatrixHelper<Scalar>& h, const typename MatrixHelper<Scalar>::DeepCopy& d)    
-    :   Base(MatrixCommitment::RowVector(), h,d) { }
-    /// @}
-
-	// This gives the resulting rowvector type when (r(i) op P) is applied to each element.
-    // It will have element types which are the regular composite result of ELT op P.
-    template <class P> struct EltResult { 
-        typedef RowVectorBase<typename CNT<ELT>::template Result<P>::Mul> Mul;
-        typedef RowVectorBase<typename CNT<ELT>::template Result<P>::Dvd> Dvd;
-        typedef RowVectorBase<typename CNT<ELT>::template Result<P>::Add> Add;
-        typedef RowVectorBase<typename CNT<ELT>::template Result<P>::Sub> Sub;
-    };
-
-    /// Copy assignment is deep copy but behavior depends on type of lhs: if view, rhs
-    /// must match. If owner, we reallocate and copy rhs.
-    RowVectorBase& operator=(const RowVectorBase& b) {
-        Base::operator=(b); return *this;
-    }
-
-    // default destructor
-
-    RowVectorBase& operator*=(const StdNumber& t)     {Base::operator*=(t); return *this;}
-    RowVectorBase& operator/=(const StdNumber& t)     {Base::operator/=(t); return *this;}
-    RowVectorBase& operator+=(const RowVectorBase& r) {Base::operator+=(r); return *this;}
-    RowVectorBase& operator-=(const RowVectorBase& r) {Base::operator-=(r); return *this;}  
-
-    template <class EE> RowVectorBase& operator=(const RowVectorBase<EE>& b) 
-      { Base::operator=(b);  return *this; } 
-    template <class EE> RowVectorBase& operator+=(const RowVectorBase<EE>& b) 
-      { Base::operator+=(b); return *this; } 
-    template <class EE> RowVectorBase& operator-=(const RowVectorBase<EE>& b) 
-      { Base::operator-=(b); return *this; } 
-
-    // default destructor
- 
-    /// Fill current allocation with copies of element. Note that this is not the 
-    /// same behavior as assignment for Matrices, where only the diagonal is set (and
-    /// everything else is set to zero.)
-    RowVectorBase& operator=(const ELT& t) { Base::setTo(t); return *this; } 
-
-    /// There's only one row here so it's a bit wierd to use colScale rather than
-    /// elementwiseMultiply, but there's nothing really wrong with it. Using rowScale
-    /// would be really wacky since it is the same as a scalar multiply. We won't support
-    /// rowScale here except through inheritance where it will not be much use.
-    template <class EE> RowVectorBase& colScaleInPlace(const VectorBase<EE>& v)
-	{ Base::template colScaleInPlace<EE>(v); return *this; }
-    template <class EE> inline void colScale(const VectorBase<EE>& v, typename EltResult<EE>::Mul& out) const
-	{ return Base::template colScale<EE>(v,out); }
-    template <class EE> inline typename EltResult<EE>::Mul colScale(const VectorBase<EE>& v) const
-	{ typename EltResult<EE>::Mul out(ncol()); Base::template colScale<EE>(v,out); return out; }
-
-
-    // elementwise multiply
-    template <class EE> RowVectorBase& elementwiseMultiplyInPlace(const RowVectorBase<EE>& r)
-	{ Base::template elementwiseMultiplyInPlace<EE>(r); return *this; }
-    template <class EE> inline void elementwiseMultiply(const RowVectorBase<EE>& v, typename EltResult<EE>::Mul& out) const
-	{ Base::template elementwiseMultiply<EE>(v,out); }
-    template <class EE> inline typename EltResult<EE>::Mul elementwiseMultiply(const RowVectorBase<EE>& v) const
-	{ typename EltResult<EE>::Mul out(nrow()); Base::template elementwiseMultiply<EE>(v,out); return out; }
-
-    // elementwise multiply from left
-    template <class EE> RowVectorBase& elementwiseMultiplyFromLeftInPlace(const RowVectorBase<EE>& r)
-	{ Base::template elementwiseMultiplyFromLeftInPlace<EE>(r); return *this; }
-    template <class EE> inline void 
-    elementwiseMultiplyFromLeft(
-        const RowVectorBase<EE>& v, 
-        typename RowVectorBase<EE>::template EltResult<ELT>::Mul& out) const
-	{ 
-        Base::template elementwiseMultiplyFromLeft<EE>(v,out);
-    }
-    template <class EE> inline 
-    typename RowVectorBase<EE>::template EltResult<ELT>::Mul 
-    elementwiseMultiplyFromLeft(const RowVectorBase<EE>& v) const {
-        typename RowVectorBase<EE>::template EltResult<ELT>::Mul out(nrow()); 
-        Base::template elementwiseMultiplyFromLeft<EE>(v,out); 
-        return out;
-    }
-
-    // elementwise divide
-    template <class EE> RowVectorBase& elementwiseDivideInPlace(const RowVectorBase<EE>& r)
-	{ Base::template elementwiseDivideInPlace<EE>(r); return *this; }
-    template <class EE> inline void elementwiseDivide(const RowVectorBase<EE>& v, typename EltResult<EE>::Dvd& out) const
-	{ Base::template elementwiseDivide<EE>(v,out); }
-    template <class EE> inline typename EltResult<EE>::Dvd elementwiseDivide(const RowVectorBase<EE>& v) const
-	{ typename EltResult<EE>::Dvd out(nrow()); Base::template elementwiseDivide<EE>(v,out); return out; }
-
-    // elementwise divide from left
-    template <class EE> RowVectorBase& elementwiseDivideFromLeftInPlace(const RowVectorBase<EE>& r)
-	{ Base::template elementwiseDivideFromLeftInPlace<EE>(r); return *this; }
-    template <class EE> inline void 
-    elementwiseDivideFromLeft
-       (const RowVectorBase<EE>& v, 
-        typename RowVectorBase<EE>::template EltResult<ELT>::Dvd& out) const { 
-        Base::template elementwiseDivideFromLeft<EE>(v,out);
-    }
-    template <class EE> inline 
-    typename RowVectorBase<EE>::template EltResult<ELT>::Dvd 
-    elementwiseDivideFromLeft(const RowVectorBase<EE>& v) const	{ 
-        typename RowVectorBase<EE>::template EltResult<ELT>::Dvd out(nrow()); 
-        Base::template elementwiseDivideFromLeft<EE>(v,out); 
-        return out;
-    }
-
-    // Implicit conversions are allowed to RowVector or Matrix, but not to Vector.   
-    operator const RowVector_<ELT>&()     const {return *reinterpret_cast<const RowVector_<ELT>*>(this);}
-    operator       RowVector_<ELT>&()           {return *reinterpret_cast<      RowVector_<ELT>*>(this);}
-    operator const RowVectorView_<ELT>&() const {return *reinterpret_cast<const RowVectorView_<ELT>*>(this);}
-    operator       RowVectorView_<ELT>&()       {return *reinterpret_cast<      RowVectorView_<ELT>*>(this);}
-    
-    operator const Matrix_<ELT>&()     const {return *reinterpret_cast<const Matrix_<ELT>*>(this);}
-    operator       Matrix_<ELT>&()           {return *reinterpret_cast<      Matrix_<ELT>*>(this);} 
-    operator const MatrixView_<ELT>&() const {return *reinterpret_cast<const MatrixView_<ELT>*>(this);}
-    operator       MatrixView_<ELT>&()       {return *reinterpret_cast<      MatrixView_<ELT>*>(this);} 
-    
-
-    // size() for RowVectors is Base::nelt() but returns int instead of ptrdiff_t.
-    int size() const { 
-	    assert(Base::nelt() <= (ptrdiff_t)std::numeric_limits<int>::max()); 
-	    assert(Base::nrow()==1);
-	    return (int)Base::nelt();
-    }
-    int       nrow() const {assert(Base::nrow()==1); return Base::nrow();}
-    int       ncol() const {assert(Base::nrow()==1); return Base::ncol();}
-    ptrdiff_t nelt() const {assert(Base::nrow()==1); return Base::nelt();}
-
-    // Override MatrixBase operators to return the right shape
-    TAbs abs() const {
-        TAbs result; Base::abs(result); return result;
-    }
-
-    // Override MatrixBase indexing operators          
-    const ELT& operator[](int j) const {return *reinterpret_cast<const ELT*>(Base::getHelper().getElt(j));}
-    ELT&       operator[](int j)       {return *reinterpret_cast<ELT*>      (Base::updHelper().updElt(j));}
-    const ELT& operator()(int j) const {return *reinterpret_cast<const ELT*>(Base::getHelper().getElt(j));}
-    ELT&       operator()(int j)       {return *reinterpret_cast<ELT*>      (Base::updHelper().updElt(j));}
-         
-    // Block (contiguous subvector) creation      
-    RowVectorView_<ELT> operator()(int j, int n) const {return Base::operator()(0,j,1,n).getAsRowVectorView();}
-    RowVectorView_<ELT> operator()(int j, int n)       {return Base::operator()(0,j,1,n).updAsRowVectorView();}
-
-    // Indexed view creation (arbitrary subvector). Indices must be monotonically increasing.
-    RowVectorView_<ELT> index(const Array_<int>& indices) const {
-        MatrixHelper<Scalar> h(Base::getHelper().getCharacterCommitment(), Base::getHelper(), indices);
-        return RowVectorView_<ELT>(h);
-    }
-    RowVectorView_<ELT> updIndex(const Array_<int>& indices) {
-        MatrixHelper<Scalar> h(Base::getHelper().getCharacterCommitment(), Base::updHelper(), indices);
-        return RowVectorView_<ELT>(h);
-    }
-
-    RowVectorView_<ELT> operator()(const Array_<int>& indices) const {return index(indices);}
-    RowVectorView_<ELT> operator()(const Array_<int>& indices)       {return updIndex(indices);}
- 
-    // Hermitian transpose.
-    THerm transpose() const {return Base::transpose().getAsVectorView();}
-    THerm updTranspose()    {return Base::updTranspose().updAsVectorView();}
-
-    THerm operator~() const {return transpose();}
-    THerm operator~()       {return updTranspose();}
-
-    const RowVectorBase& operator+() const {return *this; }
-
-    // Negation
-
-    const TNeg& negate()    const {return *reinterpret_cast<const TNeg*>(this); }
-    TNeg&       updNegate()       {return *reinterpret_cast<TNeg*>(this); }
-
-    const TNeg& operator-() const {return negate();}
-    TNeg&       operator-()       {return updNegate();}
-
-    RowVectorBase& resize(int n)     {Base::resize(1,n); return *this;}
-    RowVectorBase& resizeKeep(int n) {Base::resizeKeep(1,n); return *this;}
-
-    //TODO: this is not re-locking the number of rows at 1.
-    void clear() {Base::clear(); Base::resize(1,0);}
-
-    ELT sum() const {ELT s; Base::getHelper().sum(reinterpret_cast<Scalar*>(&s)); return s; } // add all the elements        
-    VectorIterator<ELT, RowVectorBase<ELT> > begin() {
-        return VectorIterator<ELT, RowVectorBase<ELT> >(*this, 0);
-    }
-    VectorIterator<ELT, RowVectorBase<ELT> > end() {
-        return VectorIterator<ELT, RowVectorBase<ELT> >(*this, size());
-    }
-
-protected:
-    // Create a RowVectorBase handle using a given helper rep. 
-    explicit RowVectorBase(MatrixHelperRep<Scalar>* hrep) : Base(hrep) {}
-
-private:
-    // NO DATA MEMBERS ALLOWED
-};
-
-
-
-//  ------------------------------- MatrixView_ --------------------------------
-/// This class is identical to a Matrix_; it is used only to manage the C++ rules
-/// for when copy constructors are called by introducing a separate type to
-/// prevent certain allowed optimizations from occuring when we don't want them.
-/// Despite the name, this may be an owner if a Matrix_ is recast to a MatrixView_.
-/// However, there are no owner constructors for MatrixView_. 
-//  ----------------------------------------------------------------------------
-template <class ELT> class MatrixView_ : public MatrixBase<ELT> {
-    typedef MatrixBase<ELT>                 Base;
-    typedef typename CNT<ELT>::Scalar       S;
-    typedef typename CNT<ELT>::StdNumber    StdNumber;
-public:
-    // Default construction is suppressed.
-    // Uses default destructor.
-
-    // Create a MatrixView_ handle using a given helper rep. 
-    explicit MatrixView_(MatrixHelperRep<S>* hrep) : Base(hrep) {}
-
-    // Copy constructor is shallow. CAUTION: despite const argument, this preserves writability
-    // if it was present in the source. This is necessary to allow temporary views to be
-    // created and used as lvalues.
-    MatrixView_(const MatrixView_& m) 
-      : Base(MatrixCommitment(),
-             const_cast<MatrixHelper<S>&>(m.getHelper()), 
-             typename MatrixHelper<S>::ShallowCopy()) {}
-
-    // Copy assignment is deep but not reallocating.
-    MatrixView_& operator=(const MatrixView_& m) {
-        Base::operator=(m); return *this;
-    }
-
-    // Ask for shallow copy    
-    MatrixView_(const MatrixHelper<S>& h) : Base(MatrixCommitment(), h, typename MatrixHelper<S>::ShallowCopy()) { }
-    MatrixView_(MatrixHelper<S>&       h) : Base(MatrixCommitment(), h, typename MatrixHelper<S>::ShallowCopy()) { }
-
-    MatrixView_& operator=(const Matrix_<ELT>& v)     { Base::operator=(v); return *this; }
-    MatrixView_& operator=(const ELT& e)              { Base::operator=(e); return *this; }
-
-    template <class EE> MatrixView_& operator=(const MatrixBase<EE>& m)
-      { Base::operator=(m); return *this; }
-    template <class EE> MatrixView_& operator+=(const MatrixBase<EE>& m)
-      { Base::operator+=(m); return *this; }
-    template <class EE> MatrixView_& operator-=(const MatrixBase<EE>& m)
-      { Base::operator-=(m); return *this; }
-
-    MatrixView_& operator*=(const StdNumber& t) { Base::operator*=(t); return *this; }
-    MatrixView_& operator/=(const StdNumber& t) { Base::operator/=(t); return *this; }
-    MatrixView_& operator+=(const ELT& r)       { this->updDiag() += r; return *this; }
-    MatrixView_& operator-=(const ELT& r)       { this->updDiag() -= r; return *this; }  
-
-    operator const Matrix_<ELT>&() const { return *reinterpret_cast<const Matrix_<ELT>*>(this); }
-    operator Matrix_<ELT>&()             { return *reinterpret_cast<Matrix_<ELT>*>(this); }      
-
-private:
-    // NO DATA MEMBERS ALLOWED
-    MatrixView_(); // default constructor suppressed (what's it a view of?)
-};
-
-
-//------------------------------------ Matrix_ ---------------------------------
-/** @brief This is the matrix class intended to appear in user code for large,
-variable size matrices.
-
-@ingroup MatVecUtilities
-
-More commonly, the typedef @ref SimTK::Matrix "Matrix" is used instead; that is
-just an abbreviation for \c Matrix_<Real>.
-
-A %Matrix_ can be a fixed-size view of someone else's data, or can be a 
-resizable data owner itself.
-
-@see Mat for handling of small, fixed-size matrices with no runtime overhead.
-@see Vector_ for variable size, one-dimensional column vector.
-**/
-//------------------------------------------------------------------------------
-template <class ELT> class Matrix_ : public MatrixBase<ELT> {
-    typedef typename CNT<ELT>::Scalar       S;
-    typedef typename CNT<ELT>::Number       Number;
-    typedef typename CNT<ELT>::StdNumber    StdNumber;
-
-    typedef typename CNT<ELT>::TNeg         ENeg;
-    typedef typename CNT<ELT>::THerm        EHerm;
-
-    typedef MatrixBase<ELT>     Base;
-    typedef MatrixBase<ENeg>    BaseNeg;
-    typedef MatrixBase<EHerm>   BaseHerm;
-
-    typedef Matrix_<ELT>        T;
-    typedef MatrixView_<ELT>    TView;
-    typedef Matrix_<ENeg>       TNeg;
-
-public:
-    Matrix_() : Base() { }
-    Matrix_(const MatrixCommitment& mc) : Base(mc) {}
-
-    // Copy constructor is deep.
-    Matrix_(const Matrix_& src) : Base(src) { }
-
-    // Assignment is a deep copy and will also allow reallocation if this Matrix
-    // doesn't have a view.
-    Matrix_& operator=(const Matrix_& src) { 
-        Base::operator=(src); return *this;
-    }
-
-    // Force a deep copy of the view or whatever this is.
-    // Note that this is an implicit conversion.
-    Matrix_(const Base& v) : Base(v) {}   // e.g., MatrixView
-
-    // Allow implicit conversion from a source matrix that
-    // has a negated version of ELT.
-    Matrix_(const BaseNeg& v) : Base(v) {}
-
-    // TODO: implicit conversion from conjugate. This is trickier
-    // since real elements are their own conjugate so you'll get
-    // duplicate methods defined from Matrix_(BaseHerm) and Matrix_(Base).
-
-    Matrix_(int m, int n) : Base(MatrixCommitment(), m, n) {}
-
-    Matrix_(int m, int n, const ELT* cppInitialValuesByRow) 
-    :   Base(MatrixCommitment(), m, n, cppInitialValuesByRow) {}
-    Matrix_(int m, int n, const ELT& initialValue) 
-    :   Base(MatrixCommitment(), m, n, initialValue) {}
-    
-    Matrix_(int m, int n, int leadingDim, const S* data) // read only
-    :   Base(MatrixCommitment(), MatrixCharacter::LapackFull(m,n), 
-             leadingDim, data) {}
-    Matrix_(int m, int n, int leadingDim, S* data) // writable
-    :   Base(MatrixCommitment(), MatrixCharacter::LapackFull(m,n), 
-             leadingDim, data) {}
-    
-    /// Convert a Mat to a Matrix_.
-    template <int M, int N, int CS, int RS>
-    explicit Matrix_(const Mat<M,N,ELT,CS,RS>& mat)
-    :   Base(MatrixCommitment(), M, N)
-    {   for (int i = 0; i < M; ++i)
-            for (int j = 0; j < N; ++j)
-                this->updElt(i, j) = mat(i, j); }
-
-    Matrix_& operator=(const ELT& v) { Base::operator=(v); return *this; }
-
-    template <class EE> Matrix_& operator=(const MatrixBase<EE>& m)
-      { Base::operator=(m); return*this; }
-    template <class EE> Matrix_& operator+=(const MatrixBase<EE>& m)
-      { Base::operator+=(m); return*this; }
-    template <class EE> Matrix_& operator-=(const MatrixBase<EE>& m)
-      { Base::operator-=(m); return*this; }
-
-    Matrix_& operator*=(const StdNumber& t) { Base::operator*=(t); return *this; }
-    Matrix_& operator/=(const StdNumber& t) { Base::operator/=(t); return *this; }
-    Matrix_& operator+=(const ELT& r)       { this->updDiag() += r; return *this; }
-    Matrix_& operator-=(const ELT& r)       { this->updDiag() -= r; return *this; }  
-
-    const TNeg& negate()    const {return *reinterpret_cast<const TNeg*>(this); }
-    TNeg&       updNegate()       {return *reinterpret_cast<TNeg*>(this); }
-
-    const TNeg& operator-() const {return negate();}
-    TNeg&       operator-()       {return updNegate();}
-   
-    // Functions to be used for Scripting in MATLAB and languages that do not support operator overloading
-    /** toString() returns a string representation of the Matrix_. Please refer to operator<< for details. **/
-    std::string toString() const {
-		std::stringstream stream;
-	    stream <<  (*this) ;
-		return stream.str(); 
-    }
-    /** Variant of indexing operator that's scripting friendly to get entry (i, j) **/
-    const ELT& get(int i,int j) const { return this->getElt(i,j); }
-    /** Variant of indexing operator that's scripting friendly to set entry (i, j) **/
-    void       set(int i,int j, const ELT& value)       { this->updElt(i,j)=value; }
-
-private:
-    // NO DATA MEMBERS ALLOWED
-};
-
-
-
-//  -------------------------------- VectorView_ -------------------------------
-/// This class is identical to a Vector_; it is used only to manage the C++ rules
-/// for when copy constructors are called by introducing a separate type to
-/// prevent certain allowed optimizations from occuring when we don't want them.
-/// Despite the name, this may be an owner if a Vector_ is recast to a VectorView_.
-/// However, there are no owner constructors for VectorView_. 
-//  ----------------------------------------------------------------------------
-template <class ELT> class VectorView_ : public VectorBase<ELT> {
-    typedef VectorBase<ELT>                             Base;
-    typedef typename CNT<ELT>::Scalar                   S;
-    typedef typename CNT<ELT>::Number                   Number;
-    typedef typename CNT<ELT>::StdNumber                StdNumber;
-    typedef VectorView_<ELT>                            T;
-    typedef VectorView_< typename CNT<ELT>::TNeg >      TNeg;
-    typedef RowVectorView_< typename CNT<ELT>::THerm >  THerm;
-public:
-    // Default construction is suppressed.
-    // Uses default destructor.
-
-    // Create a VectorView_ handle using a given helper rep. 
-    explicit VectorView_(MatrixHelperRep<S>* hrep) : Base(hrep) {}
-
-    // Copy constructor is shallow. CAUTION: despite const argument, this preserves writability
-    // if it was present in the source. This is necessary to allow temporary views to be
-    // created and used as lvalues.
-    VectorView_(const VectorView_& v) 
-      : Base(const_cast<MatrixHelper<S>&>(v.getHelper()), typename MatrixHelper<S>::ShallowCopy()) { }
-
-    // Copy assignment is deep but not reallocating.
-    VectorView_& operator=(const VectorView_& v) {
-        Base::operator=(v); return *this;
-    }
-
-    // Ask for shallow copy    
-    explicit VectorView_(const MatrixHelper<S>& h) : Base(h, typename MatrixHelper<S>::ShallowCopy()) { }
-    explicit VectorView_(MatrixHelper<S>&       h) : Base(h, typename MatrixHelper<S>::ShallowCopy()) { }
-    
-    VectorView_& operator=(const Base& b) { Base::operator=(b); return *this; }
-
-    VectorView_& operator=(const ELT& v) { Base::operator=(v); return *this; } 
-
-    template <class EE> VectorView_& operator=(const VectorBase<EE>& m)
-      { Base::operator=(m); return*this; }
-    template <class EE> VectorView_& operator+=(const VectorBase<EE>& m)
-      { Base::operator+=(m); return*this; }
-    template <class EE> VectorView_& operator-=(const VectorBase<EE>& m)
-      { Base::operator-=(m); return*this; }
-
-    VectorView_& operator*=(const StdNumber& t) { Base::operator*=(t); return *this; }
-    VectorView_& operator/=(const StdNumber& t) { Base::operator/=(t); return *this; }
-    VectorView_& operator+=(const ELT& b) { this->elementwiseAddScalarInPlace(b); return *this; }
-    VectorView_& operator-=(const ELT& b) { this->elementwiseSubtractScalarInPlace(b); return *this; }
-
-private:
-    // NO DATA MEMBERS ALLOWED
-    VectorView_(); // default construction suppressed (what's it a View of?)
-};
-
-
-//------------------------------------ Vector_ ---------------------------------
-/** @brief This is the vector class intended to appear in user code for large,
-variable size column vectors.
-
-@ingroup MatVecUtilities
-
-More commonly, the typedef @ref SimTK::Vector "Vector" is used instead; that is
-just an abbreviation for \c Vector_<Real>.
-
-A %Vector_ can be a fixed-size view of someone else's data, or can be a 
-resizable data owner itself, although of course it will always have just one 
-column.
-
-@see Vec for handling of small, fixed-size vectors with no runtime overhead.
-@see Matrix_ for variable %size, two-dimensional matrix.
-**/
-template <class ELT> class Vector_ : public VectorBase<ELT> {
-    typedef typename CNT<ELT>::Scalar       S;
-    typedef typename CNT<ELT>::Number       Number;
-    typedef typename CNT<ELT>::StdNumber    StdNumber;
-    typedef typename CNT<ELT>::TNeg         ENeg;
-    typedef VectorBase<ELT>                 Base;
-    typedef VectorBase<ENeg>                BaseNeg;
-public:
-    // Uses default destructor.
-
-    /** @name                  Owner constructors
-    These constructors create a Vector_ object that allocates new data on the
-    heap and serves as the owner of that data. Any existing data that is 
-    supplied in these constructors is \e copied, not \e shared. **/
-    /**@{**/
-
-    /** Default constructor creates a 0x1, reallocatable vector. **/
-    Vector_() : Base() {}
-
-    /** Copy constructor is deep, that is the source %Vector_ is copied, not
-    referenced. **/
-    Vector_(const Vector_& src) : Base(src) {}
-
-    /** This copy constructor serves as an implicit conversion from objects of
-    the base class type (for example, a VectorView_<ELT>), to objects of this 
-    Vector_<ELT> type. Note that the source object is copied, not 
-    referenced. **/
-    Vector_(const Base& src) : Base(src) {}    // e.g., VectorView
-    /** This copy constructor serves as an implicit conversion from objects of
-    the base class but with negated elements, to objects of this Vector_<ELT>
-    type.  Note that the source object is copied, not referenced. **/
-    Vector_(const BaseNeg& src) : Base(src) {}
-
-    /** Construct a %Vector_ with a given preallocated size, but with 
-    uninitialized values. In Debug builds the elements will be initialized to
-    NaN as a debugging aid, but in Release (optimized) builds they will be 
-    uninitialized garbage so that we don't waste time setting them. **/
-    explicit Vector_(int m) : Base(m) { }
-    /** Construct an owner %Vector_ of a given size \a m and initialize it from 
-    a C++ array of \a m ELT values pointed to by \a cppInitialValues. Note that 
-    there is no way to check that the correct number of elements has been
-    provided; make sure you have supplied enough of them. **/
-    Vector_(int m, const ELT* cppInitialValues) : Base(m, cppInitialValues) {}
-    /** Construct an owner %Vector_ of a given size \a m and initialize all the
-    elements to the given ELT value \a initialValue. **/
-    Vector_(int m, const ELT& initialValue) : Base(m, initialValue) {}
-
-    /** This constructor creates a new owner %Vector_ that is the same length
-    and has a copy of the contents of a given fixed-size Vec. **/
-    template <int M>
-    explicit Vector_(const Vec<M,ELT>& v) : Base(M) {
-        for (int i = 0; i < M; ++i)
-            this->updElt(i, 0) = v(i);
-    }
-    /**@}**/
-
-    /** @name                  View constructors
-    These constructors create a %Vector_ object that is a view of existing
-    data that is owned by some other object. Any existing data that is
-    supplied in these constructors is \e shared, not \e copied, so writing
-    to the created %Vector_ view modifies the original data. **/
-    /**@{**/
-
-    /** Construct a %Vector_ which shares read-only, borrowed space with assumed
-    element-to-element stride equal to the C++ element spacing. Last parameter
-    is a required dummy to avoid overload conflicts when ELT=S; pass it as
-    "true". **/
-    Vector_(int m, const S* cppData, bool)
-        : Base(m, Base::CppNScalarsPerElement, cppData) {}
-    /** Construct a %Vector_ which shares writable, borrowed space with assumed
-    element-to-element stride equal to the C++ element spacing. Last parameter
-    is a required dummy to avoid overload conflicts when ELT=S; pass it as
-    "true". **/
-    Vector_(int m, S* cppData, bool)
-        : Base(m, Base::CppNScalarsPerElement, cppData) {}
-
-    /** Borrowed-space read-only construction with explicit stride supplied as
-    "number of scalars between elements". Last parameter is a dummy to avoid
-    overload conflicts; pass it as "true". **/
-    Vector_(int m, int stride, const S* data, bool)
-        : Base(m, stride, data) {}
-    /** Borrowed-space writable construction with explicit stride supplied as
-    "number of scalars between elements". Last parameter is a dummy to avoid
-    overload conflicts; pass it as "true". **/
-    Vector_(int m, int stride, S* data, bool)
-        : Base(m, stride, data) {}
-    /**@}**/
-
-    /** @name              Assignment operators **/
-    /**@{**/
-    /** Copy assignment is deep and can be reallocating if this %Vector_ is an
-    owner. Otherwise the source is copied into the destination view. **/
-    Vector_& operator=(const Vector_& src)
-    {   Base::operator=(src); return*this; }
-
-    /** Like copy assignment but the source can be any object of the base
-    type, even with a different element type \a EE as long as that element
-    type is assignment-compatible to an element of our type \a ELT. **/
-    template <class EE> Vector_& operator=(const VectorBase<EE>& src)
-    {   Base::operator=(src); return*this; }
-
-    /** Set all elements of this %Vector_ to the same value \a v. **/
-    Vector_& operator=(const ELT& v) { Base::operator=(v); return *this; }
-
-    /** Add in a conforming vector of the base type even if it has a different
-    element type \a EE, provided that element type is add-compatible with
-    our element type \a ELT, meaning that ELT+=EE is allowed. **/
-    template <class EE> Vector_& operator+=(const VectorBase<EE>& m)
-    {   Base::operator+=(m); return*this; }
-    /** In-place subtract of a conforming vector of the base type even if it has
-    a different element type \a EE, provided that element type is 
-    subtract-compatible with our element type \a ELT, meaning that ELT-=EE is
-    allowed. **/
-    template <class EE> Vector_& operator-=(const VectorBase<EE>& m)
-    {   Base::operator-=(m); return*this; }
-
-    /** In-place multiply of each element of this %Vector_ by a scalar \a t.
-    Returns a reference to the now-modified %Vector_. **/
-    Vector_& operator*=(const StdNumber& t) { Base::operator*=(t); return *this; }
-    /** In-place divide of each element of this %Vector_ by a scalar \a t.
-    Returns a reference to the now-modified %Vector_. **/
-    Vector_& operator/=(const StdNumber& t) { Base::operator/=(t); return *this; }
-
-    /** In-place add to each element of this %Vector_ the same given value \a b.
-    Returns a reference to the now-modified %Vector_. This is the same 
-    operation as elementwiseAddScalarInPlace(). **/
-    Vector_& operator+=(const ELT& b) 
-    {   this->elementwiseAddScalarInPlace(b); return *this; }
-    /** In-place subtract from each element of this %Vector_ the same given 
-    value \a b. Returns a reference to the now-modified %Vector_. This is the 
-    same operation as elementwiseSubtractScalarInPlace(). **/
-    Vector_& operator-=(const ELT& b) 
-    {   this->elementwiseSubtractScalarInPlace(b); return *this; }
-    /**@}**/
-
-    /** @name             Operator equivalents for scripting
-    Functions to be used for Scripting in MATLAB and languages that do not 
-    support operator overloading. **/
-    /**@{**/
-    /** toString() returns a string representation of the %Vector_. Please refer 
-    to operator<< for details. **/
-    std::string toString() const {
-		std::stringstream stream;
-	    stream << (*this) ;
-		return stream.str(); 
-    }
-    /** Variant of operator[] that's scripting friendly to get ith element. **/
-    const ELT& get(int i) const { return (*this)[i]; }
-    /** Variant of operator[] that's scripting friendly to set ith element. **/
-    void set(int i, const ELT& value)  { (*this)[i]=value; }
-    /**@}**/
-
-private:
-    // NO DATA MEMBERS ALLOWED
-};
-
-
-
-//  ------------------------------ RowVectorView_ ------------------------------
-/// This class is identical to a RowVector_; it is used only to manage the C++ 
-/// rules for when copy constructors are called by introducing a separate type to
-/// prevent certain allowed optimizations from occuring when we don't want them.
-/// Despite the name, this may be an owner if a RowVector_ is recast to a 
-/// RowVectorView_. However, there are no owner constructors for RowVectorView_. 
-//  ----------------------------------------------------------------------------
-template <class ELT> class RowVectorView_ : public RowVectorBase<ELT> {
-    typedef RowVectorBase<ELT>                              Base;
-    typedef typename CNT<ELT>::Scalar                       S;
-    typedef typename CNT<ELT>::Number                       Number;
-    typedef typename CNT<ELT>::StdNumber                    StdNumber;
-    typedef RowVectorView_<ELT>                             T;
-    typedef RowVectorView_< typename CNT<ELT>::TNeg >       TNeg;
-    typedef VectorView_< typename CNT<ELT>::THerm >         THerm;
-public:
-    // Default construction is suppressed.
-    // Uses default destructor.
-
-    // Create a RowVectorView_ handle using a given helper rep. 
-    explicit RowVectorView_(MatrixHelperRep<S>* hrep) : Base(hrep) {}
-
-    // Copy constructor is shallow. CAUTION: despite const argument, this preserves writability
-    // if it was present in the source. This is necessary to allow temporary views to be
-    // created and used as lvalues.
-    RowVectorView_(const RowVectorView_& r) 
-      : Base(const_cast<MatrixHelper<S>&>(r.getHelper()), typename MatrixHelper<S>::ShallowCopy()) { }
-
-    // Copy assignment is deep but not reallocating.
-    RowVectorView_& operator=(const RowVectorView_& r) {
-        Base::operator=(r); return *this;
-    }
-
-    // Ask for shallow copy    
-    explicit RowVectorView_(const MatrixHelper<S>& h) : Base(h, typename MatrixHelper<S>::ShallowCopy()) { }
-    explicit RowVectorView_(MatrixHelper<S>&       h) : Base(h, typename MatrixHelper<S>::ShallowCopy()) { }
-    
-    RowVectorView_& operator=(const Base& b) { Base::operator=(b); return *this; }
-
-    RowVectorView_& operator=(const ELT& v) { Base::operator=(v); return *this; } 
-
-    template <class EE> RowVectorView_& operator=(const RowVectorBase<EE>& m)
-      { Base::operator=(m); return*this; }
-    template <class EE> RowVectorView_& operator+=(const RowVectorBase<EE>& m)
-      { Base::operator+=(m); return*this; }
-    template <class EE> RowVectorView_& operator-=(const RowVectorBase<EE>& m)
-      { Base::operator-=(m); return*this; }
-
-    RowVectorView_& operator*=(const StdNumber& t) { Base::operator*=(t); return *this; }
-    RowVectorView_& operator/=(const StdNumber& t) { Base::operator/=(t); return *this; }
-    RowVectorView_& operator+=(const ELT& b) { this->elementwiseAddScalarInPlace(b); return *this; }
-    RowVectorView_& operator-=(const ELT& b) { this->elementwiseSubtractScalarInPlace(b); return *this; }
-
-private:
-    // NO DATA MEMBERS ALLOWED
-    RowVectorView_(); // default construction suppressed (what is it a view of?)
-};
-
-
-
-//  -------------------------------- RowVector_ --------------------------------
-/// RowVectors are much less common than Vectors. However, if a Simmatrix user 
-/// wants one, this is the class intended to appear in user code. It can be a 
-/// fixed-size view of someone else's data, or can be a resizable data owner 
-/// itself, although of course it will always have just one row.
-//  ----------------------------------------------------------------------------
-template <class ELT> class RowVector_ : public RowVectorBase<ELT> {
-    typedef typename CNT<ELT>::Scalar       S;
-    typedef typename CNT<ELT>::Number       Number;
-    typedef typename CNT<ELT>::StdNumber    StdNumber;
-    typedef typename CNT<ELT>::TNeg         ENeg;
-
-    typedef RowVectorBase<ELT>              Base;
-    typedef RowVectorBase<ENeg>             BaseNeg;
-public:
-    RowVector_() : Base() {}   // 1x0 reallocatable
-    // Uses default destructor.
-
-    // Copy constructor is deep.
-    RowVector_(const RowVector_& src) : Base(src) {}
-
-    // Implicit conversions.
-    RowVector_(const Base& src) : Base(src) {}    // e.g., RowVectorView
-    RowVector_(const BaseNeg& src) : Base(src) {}  
-
-    // Copy assignment is deep and can be reallocating if this RowVector
-    // has no View.
-    RowVector_& operator=(const RowVector_& src) {
-        Base::operator=(src); return*this;
-    }
-
-
-    explicit RowVector_(int n) : Base(n) { }
-    RowVector_(int n, const ELT* cppInitialValues) : Base(n, cppInitialValues) {}
-    RowVector_(int n, const ELT& initialValue)     : Base(n, initialValue) {}
-
-    /// Construct a Vector which uses borrowed space with assumed
-    /// element-to-element stride equal to the C++ element spacing.
-    /// Last parameter is a dummy to avoid overload conflicts when ELT=S;
-    /// pass it as "true".
-    RowVector_(int n, const S* cppData, bool): Base(n, Base::CppNScalarsPerElement, cppData) {}
-    RowVector_(int n,       S* cppData, bool): Base(n, Base::CppNScalarsPerElement, cppData) {}
-
-    /// Borrowed-space construction with explicit stride supplied as
-    /// "number of scalars between elements". Last parameter is a 
-    /// dummy to avoid overload conflicts; pass it as "true".
-    RowVector_(int n, int stride, const S* data, bool) : Base(n, stride, data) {}
-    RowVector_(int n, int stride,       S* data, bool) : Base(n, stride, data) {}
-    
-    /// Convert a Row to a RowVector_.
-    template <int M>
-    explicit RowVector_(const Row<M,ELT>& v) : Base(M) {
-        for (int i = 0; i < M; ++i)
-            this->updElt(0, i) = v(i);
-    }
-
-    RowVector_& operator=(const ELT& v) { Base::operator=(v); return *this; } 
-
-    template <class EE> RowVector_& operator=(const RowVectorBase<EE>& b)
-      { Base::operator=(b); return*this; }
-    template <class EE> RowVector_& operator+=(const RowVectorBase<EE>& b)
-      { Base::operator+=(b); return*this; }
-    template <class EE> RowVector_& operator-=(const RowVectorBase<EE>& b)
-      { Base::operator-=(b); return*this; }
-
-    RowVector_& operator*=(const StdNumber& t) { Base::operator*=(t); return *this; }
-    RowVector_& operator/=(const StdNumber& t) { Base::operator/=(t); return *this; }
-    RowVector_& operator+=(const ELT& b) { this->elementwiseAddScalarInPlace(b); return *this; }
-    RowVector_& operator-=(const ELT& b) { this->elementwiseSubtractScalarInPlace(b); return *this; }
-
-private:
-    // NO DATA MEMBERS ALLOWED
-};
-
-
+namespace SimTK {
 
 //  ------------------------ MatrixBase definitions ----------------------------
 
@@ -3667,100 +1558,6 @@ typedef MatrixView_<fComplex>   fComplexMatrixView;
 /** Abbreviation for MatrixView_<std::complex<double>>. **/
 typedef MatrixView_<dComplex>   dComplexMatrixView;
 /**@}**/
-
-
-/**
- * This is an iterator for iterating over the elements of a Vector.
- */
-
-template <class ELT, class VECTOR_CLASS>
-class VectorIterator {
-public:
-    typedef ELT value_type;
-    typedef ptrdiff_t difference_type;
-    typedef ELT& reference;
-    typedef ELT* pointer;
-    typedef std::random_access_iterator_tag iterator_category;
-    VectorIterator(VECTOR_CLASS& vector, ptrdiff_t index) : vector(vector), index(index) {
-    }
-    VectorIterator(const VectorIterator& iter) : vector(iter.vector), index(iter.index) {
-    }
-    VectorIterator& operator=(const VectorIterator& iter) {
-        vector = iter.vector;
-        index = iter.index;
-        return *this;
-    }
-    ELT& operator*() {
-        assert (index >= 0 && index < vector.size());
-        return vector[(int)index];
-    }
-    ELT& operator[](ptrdiff_t i) {
-        assert (i >= 0 && i < vector.size());
-        return vector[(int)i];
-    }
-    VectorIterator operator++() {
-        assert (index < vector.size());
-        ++index;
-        return *this;
-    }
-    VectorIterator operator++(int) {
-        assert (index < vector.size());
-        VectorIterator current = *this;
-        ++index;
-        return current;
-    }
-    VectorIterator operator--() {
-        assert (index > 0);
-        --index;
-        return *this;
-    }
-    VectorIterator operator--(int) {
-        assert (index > 0);
-        VectorIterator current = *this;
-        --index;
-        return current;
-    }
-    VectorIterator operator+=(ptrdiff_t n) {
-        assert (0 <= index+n && index+n <= vector.size());
-        index += n;
-        return *this;
-    }
-    VectorIterator operator-=(ptrdiff_t n) {
-        assert (0 <= index-n && index-n <= vector.size());
-        index -= n;
-        return *this;
-    }
-    bool operator<(VectorIterator iter) const {
-        return (index < iter.index);
-    }
-    bool operator>(VectorIterator iter) const {
-        return (index > iter.index);
-    }
-    bool operator<=(VectorIterator iter) const {
-        return (index <= iter.index);
-    }
-    bool operator>=(VectorIterator iter) const {
-        return (index >= iter.index);
-    }
-    ptrdiff_t operator-(VectorIterator iter) const {
-        return (index - iter.index);
-    }
-    VectorIterator operator-(ptrdiff_t n) const {
-        return VectorIterator(vector, index-n);
-    }
-    VectorIterator operator+(ptrdiff_t n) const {
-        return VectorIterator(vector, index+n);
-    }
-    bool operator==(VectorIterator iter) const {
-        return (index == iter.index);
-    }
-    bool operator!=(VectorIterator iter) const {
-        return (index != iter.index);
-    }
-private:
-    VECTOR_CLASS& vector;
-    ptrdiff_t index;
-};
 
 } //namespace SimTK
